% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/beta_dist.R
\name{cdf.beta_dist}
\alias{cdf.beta_dist}
\title{Cumulative distribution function for a beta distribution.}
\usage{
\method{cdf}{beta_dist}(x, ...)
}
\arguments{
\item{x}{A \code{beta_dist} object.}

\item{...}{Additional arguments (not used).}
}
\value{
A function \code{function(q, log.p = FALSE, ...)} returning the
CDF (or log-CDF) at \code{q}.
}
\description{
Returns a function that evaluates the beta CDF at given points.
}
\examples{
x <- beta_dist(2, 5)
F <- cdf(x)
F(0.3)
F(0.5)
}
