% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chi_squared.R
\name{cdf.chi_squared}
\alias{cdf.chi_squared}
\title{Method for obtaining the cdf of a \code{chi_squared} object.}
\usage{
\method{cdf}{chi_squared}(x, ...)
}
\arguments{
\item{x}{The \code{chi_squared} object}

\item{...}{Additional arguments (not used)}
}
\value{
A function that computes the cdf at point(s) \code{t}
}
\description{
Method for obtaining the cdf of a \code{chi_squared} object.
}
