% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exp_dist.R
\name{cdf.exp_dist}
\alias{cdf.exp_dist}
\title{Method to obtain the cdf of an \code{exp_dist} object.}
\usage{
\method{cdf}{exp_dist}(x, ...)
}
\arguments{
\item{x}{The object to obtain the pdf of}

\item{...}{Additional arguments (not used)}
}
\value{
A function that computes the cdf of the exponential. Accepts as
input a vector \code{t} at which to compute the cdf, an input \code{rate}
denoting the failure rate of the exponential distribution, and a
logical \code{log} indicating whether to compute the log of the cdf.
By default, \code{rate} is the failure rate of object \code{x}.
}
\description{
Method to obtain the cdf of an \code{exp_dist} object.
}
