% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exponential.R
\name{cdf.exponential}
\alias{cdf.exponential}
\title{Method to obtain the cdf of an \code{exponential} object.}
\usage{
\method{cdf}{exponential}(x, ...)
}
\arguments{
\item{x}{The object to obtain the cdf of}

\item{...}{Additional arguments (not used)}
}
\value{
A function \code{function(q, lower.tail = TRUE, log.p = FALSE, ...)}
that computes the cdf (or log-cdf) of the exponential distribution.
}
\description{
Method to obtain the cdf of an \code{exponential} object.
}
\examples{
x <- exponential(rate = 1)
F <- cdf(x)
F(1)
F(2)
}
