% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lognormal.R
\name{cdf.lognormal}
\alias{cdf.lognormal}
\title{Cumulative distribution function for a log-normal distribution.}
\usage{
\method{cdf}{lognormal}(x, ...)
}
\arguments{
\item{x}{A \code{lognormal} object.}

\item{...}{Additional arguments (not used).}
}
\value{
A function \code{function(q, log.p = FALSE, ...)} returning the
CDF (or log-CDF) at \code{q}.
}
\description{
Returns a function that evaluates the log-normal CDF at given points.
}
\examples{
x <- lognormal(0, 1)
F <- cdf(x)
F(1)
F(2)
}
