% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/normal.R
\name{cdf.normal}
\alias{cdf.normal}
\title{Method for obtaining the cdf of an \code{normal} object.}
\usage{
\method{cdf}{normal}(x, ...)
}
\arguments{
\item{x}{The object to obtain the cdf of}

\item{...}{Additional arguments to pass (not used)}
}
\value{
A function \code{function(q, lower.tail = TRUE, log.p = FALSE, ...)}
that computes the cdf (or log-cdf) of the normal distribution at \code{q}.
}
\description{
Method for obtaining the cdf of an \code{normal} object.
}
\examples{
x <- normal(0, 1)
F <- cdf(x)
F(0)
F(1.96)
}
