% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/poisson_dist.R
\name{cdf.poisson_dist}
\alias{cdf.poisson_dist}
\title{Cumulative distribution function for a Poisson distribution.}
\usage{
\method{cdf}{poisson_dist}(x, ...)
}
\arguments{
\item{x}{A \code{poisson_dist} object.}

\item{...}{Additional arguments (not used).}
}
\value{
A function \code{function(q, log.p = FALSE, ...)} returning the
CDF (or log-CDF) at \code{q}.
}
\description{
Returns a function that evaluates the Poisson CDF at given points.
}
\examples{
x <- poisson_dist(5)
F <- cdf(x)
F(5)
F(10)
}
