% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/uniform_dist.R
\name{cdf.uniform_dist}
\alias{cdf.uniform_dist}
\title{Cumulative distribution function for a uniform distribution.}
\usage{
\method{cdf}{uniform_dist}(x, ...)
}
\arguments{
\item{x}{A \code{uniform_dist} object.}

\item{...}{Additional arguments (not used).}
}
\value{
A function \code{function(q, log.p = FALSE, ...)} returning the
CDF (or log-CDF) at \code{q}.
}
\description{
Returns a function that evaluates the uniform CDF at given points.
}
\examples{
x <- uniform_dist(0, 10)
F <- cdf(x)
F(5)
F(10)
}
