% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chi_squared.R
\name{chi_squared}
\alias{chi_squared}
\title{Construct a chi-squared distribution object.}
\usage{
chi_squared(df)
}
\arguments{
\item{df}{Degrees of freedom (positive scalar)}
}
\value{
A \code{chi_squared} object
}
\description{
Construct a chi-squared distribution object.
}
