% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dist.R
\name{conditional.dist}
\alias{conditional.dist}
\title{Method for obtaining the condition distribution, \code{x | P(x)}, of
\code{dist} object \code{x}.}
\usage{
\method{conditional}{dist}(x, P, n = 10000L, ...)
}
\arguments{
\item{x}{The distribution object.}

\item{P}{The predicate function to condition the distribution on}

\item{n}{The number of samples to generate for the MC estimate of the
conditional distribution x | P. Defaults to 10000.}

\item{...}{additional arguments to pass into \code{P}.}
}
\description{
We just sample from \code{x} and place the sample in \code{empirical_dist} and
then condition on it.
}
