% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interval.R
\name{contains.interval}
\alias{contains.interval}
\title{Determine if a value is contained in the interval.}
\usage{
\method{contains}{interval}(object, x)
}
\arguments{
\item{object}{An interval object.}

\item{x}{A vector of values.}
}
\description{
Determine if a value is contained in the interval.
}
