% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/beta_dist.R
\name{density.beta_dist}
\alias{density.beta_dist}
\title{Probability density function for a beta distribution.}
\usage{
\method{density}{beta_dist}(x, ...)
}
\arguments{
\item{x}{A \code{beta_dist} object.}

\item{...}{Additional arguments (not used).}
}
\value{
A function \code{function(t, log = FALSE, ...)} returning the
density (or log-density) at \code{t}.
}
\description{
Returns a function that evaluates the beta PDF at given points.
}
\examples{
x <- beta_dist(2, 5)
f <- density(x)
f(0.3)
f(0.5)
}
