% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chi_squared.R
\name{density.chi_squared}
\alias{density.chi_squared}
\title{Method for obtaining the density (pdf) of a \code{chi_squared} object.}
\usage{
\method{density}{chi_squared}(x, ...)
}
\arguments{
\item{x}{The \code{chi_squared} object}

\item{...}{Additional arguments (not used)}
}
\value{
A function that computes the pdf at point(s) \code{t}
}
\description{
Method for obtaining the density (pdf) of a \code{chi_squared} object.
}
\examples{
x <- chi_squared(5)
f <- density(x)
f(5)
f(10)
}
