% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/edist.R
\name{density.edist}
\alias{density.edist}
\title{Density for expression distributions.}
\usage{
\method{density}{edist}(x, ...)
}
\arguments{
\item{x}{An \code{edist} object.}

\item{...}{Additional arguments forwarded to \code{density.empirical_dist}.}
}
\value{
A function computing the empirical density (PMF).
}
\description{
Falls back to \code{\link{realize}} and delegates to
\code{\link{density.empirical_dist}}.
}
