% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exponential.R
\name{density.exponential}
\alias{density.exponential}
\title{Method to obtain the pdf of an \code{exponential} object.}
\usage{
\method{density}{exponential}(x, ...)
}
\arguments{
\item{x}{The object to obtain the pdf of}

\item{...}{Additional arguments (not used)}
}
\value{
A function \code{function(t, log = FALSE, ...)} that computes
the pdf (or log-pdf) of the exponential distribution at \code{t}.
}
\description{
Method to obtain the pdf of an \code{exponential} object.
}
\examples{
x <- exponential(rate = 2)
f <- density(x)
f(0)
f(1)
}
