% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gamma_dist.R
\name{density.gamma_dist}
\alias{density.gamma_dist}
\title{Method for obtaining the density (pdf) of a \code{gamma_dist} object.}
\usage{
\method{density}{gamma_dist}(x, ...)
}
\arguments{
\item{x}{The \code{gamma_dist} object}

\item{...}{Additional arguments (not used)}
}
\value{
A function that computes the pdf at point(s) \code{t}
}
\description{
Method for obtaining the density (pdf) of a \code{gamma_dist} object.
}
\examples{
x <- gamma_dist(shape = 2, rate = 1)
f <- density(x)
f(1)
f(2)
}
