% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lognormal.R
\name{density.lognormal}
\alias{density.lognormal}
\title{Probability density function for a log-normal distribution.}
\usage{
\method{density}{lognormal}(x, ...)
}
\arguments{
\item{x}{A \code{lognormal} object.}

\item{...}{Additional arguments (not used).}
}
\value{
A function \code{function(t, log = FALSE, ...)} returning the
density (or log-density) at \code{t}.
}
\description{
Returns a function that evaluates the log-normal PDF at given points.
}
