% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mvn.R
\name{density.mvn}
\alias{density.mvn}
\title{Function generator for obtaining the pdf of an \code{mvn} object (multivariate
normal).}
\usage{
\method{density}{mvn}(x, ...)
}
\arguments{
\item{x}{The \code{mvn} (S3) object to obtain the pdf (density) of}

\item{...}{Additional arguments passed to \code{dmvnorm} on every call.}
}
\value{
A function \code{function(obs, log = FALSE, ...)} that computes the
pdf (or log-pdf) of the multivariate normal distribution.
}
\description{
Function generator for obtaining the pdf of an \code{mvn} object (multivariate
normal).
}
\examples{
X <- mvn(c(0, 0), diag(2))
f <- density(X)
f(c(0, 0))
f(c(1, 1))
}
