% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/normal.R
\name{density.normal}
\alias{density.normal}
\title{Method for obtaining the pdf of an \code{normal} object.}
\usage{
\method{density}{normal}(x, ...)
}
\arguments{
\item{x}{The object to obtain the pdf of}

\item{...}{Additional arguments to pass (not used)}
}
\value{
A function \code{function(t, log = FALSE, ...)} that computes the
pdf (or log-pdf) of the normal distribution at \code{t}.
}
\description{
Method for obtaining the pdf of an \code{normal} object.
}
\examples{
x <- normal(0, 1)
f <- density(x)
f(0)
f(1)
}
