% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/poisson_dist.R
\name{density.poisson_dist}
\alias{density.poisson_dist}
\title{Probability mass function for a Poisson distribution.}
\usage{
\method{density}{poisson_dist}(x, ...)
}
\arguments{
\item{x}{A \code{poisson_dist} object.}

\item{...}{Additional arguments (not used).}
}
\value{
A function \code{function(k, log = FALSE, ...)} returning the
probability mass (or log-probability) at \code{k}.
}
\description{
Returns a function that evaluates the Poisson PMF at given points.
}
