% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/uniform_dist.R
\name{density.uniform_dist}
\alias{density.uniform_dist}
\title{Probability density function for a uniform distribution.}
\usage{
\method{density}{uniform_dist}(x, ...)
}
\arguments{
\item{x}{A \code{uniform_dist} object.}

\item{...}{Additional arguments (not used).}
}
\value{
A function \code{function(t, log = FALSE, ...)} returning the
density (or log-density) at \code{t}.
}
\description{
Returns a function that evaluates the uniform PDF at given points.
}
