% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/beta_dist.R
\name{dim.beta_dist}
\alias{dim.beta_dist}
\title{Dimension of a beta distribution (always 1).}
\usage{
\method{dim}{beta_dist}(x)
}
\arguments{
\item{x}{A \code{beta_dist} object.}
}
\value{
\code{1}.
}
\description{
Dimension of a beta distribution (always 1).
}
