% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chi_squared.R
\name{dim.chi_squared}
\alias{dim.chi_squared}
\title{Retrieve the dimension of a \code{chi_squared} object.}
\usage{
\method{dim}{chi_squared}(x)
}
\arguments{
\item{x}{The \code{chi_squared} object}
}
\value{
1 (univariate)
}
\description{
Retrieve the dimension of a \code{chi_squared} object.
}
