% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/countable_set.R
\name{dim.countable_set}
\alias{dim.countable_set}
\title{Get the dimension of a countable set.}
\usage{
\method{dim}{countable_set}(x)
}
\arguments{
\item{x}{A \code{countable_set} object.}
}
\value{
\code{1} (always univariate).
}
\description{
Get the dimension of a countable set.
}
