% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/edist.R
\name{dim.edist}
\alias{dim.edist}
\title{Method for obtaining the dimension of an \code{edist} object.}
\usage{
\method{dim}{edist}(x)
}
\arguments{
\item{x}{The \code{edist} object.}
}
\value{
Integer; the number of dimensions.
}
\description{
Determines the dimension by drawing a single sample and checking
whether it is a matrix (multivariate) or scalar (univariate).
}
\examples{
z <- normal(0, 1) * exponential(1)
dim(z)
}
