% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/empirical_dist.R
\name{dim.empirical_dist}
\alias{dim.empirical_dist}
\title{Method for obtaining the dimension of a \code{empirical_dist} object.}
\usage{
\method{dim}{empirical_dist}(x)
}
\arguments{
\item{x}{The object to obtain the dimension of.}
}
\description{
Method for obtaining the dimension of a \code{empirical_dist} object.
}
