% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exponential.R
\name{dim.exponential}
\alias{dim.exponential}
\title{Method to obtain the dimension of an \code{exponential} object.}
\usage{
\method{dim}{exponential}(x)
}
\arguments{
\item{x}{The \code{exponential} object to obtain the dimension of}
}
\value{
The dimension of the \code{exponential} object
}
\description{
Method to obtain the dimension of an \code{exponential} object.
}
