% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/finite_set.R
\name{dim.finite_set}
\alias{dim.finite_set}
\title{Return the dimension of the finite set.}
\usage{
\method{dim}{finite_set}(x)
}
\arguments{
\item{x}{A finite set.}
}
\value{
Integer; the dimension of the set.
}
\description{
Return the dimension of the finite set.
}
\examples{
fs <- finite_set$new(c(1, 3, 5, 7))
dim(fs) # 1
}
