% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gamma_dist.R
\name{dim.gamma_dist}
\alias{dim.gamma_dist}
\title{Retrieve the dimension of a \code{gamma_dist} object.}
\usage{
\method{dim}{gamma_dist}(x)
}
\arguments{
\item{x}{The \code{gamma_dist} object}
}
\value{
1 (univariate)
}
\description{
Retrieve the dimension of a \code{gamma_dist} object.
}
