% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lognormal.R
\name{dim.lognormal}
\alias{dim.lognormal}
\title{Dimension of a log-normal distribution (always 1).}
\usage{
\method{dim}{lognormal}(x)
}
\arguments{
\item{x}{A \code{lognormal} object.}
}
\value{
\code{1}.
}
\description{
Dimension of a log-normal distribution (always 1).
}
