% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mixture.R
\name{dim.mixture}
\alias{dim.mixture}
\title{Dimension of a mixture distribution.}
\usage{
\method{dim}{mixture}(x)
}
\arguments{
\item{x}{A \code{mixture} object.}
}
\value{
The dimension of the distribution.
}
\description{
Returns the dimension of the first component (all components are
assumed to have the same dimension).
}
\examples{
m <- mixture(list(normal(0, 1), normal(5, 1)), c(0.5, 0.5))
dim(m)
}
