% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mvn.R
\name{dim.mvn}
\alias{dim.mvn}
\title{Method for obtaining the dimension of an \code{mvn} object.}
\usage{
\method{dim}{mvn}(x)
}
\arguments{
\item{x}{The object to obtain the dimension of}
}
\value{
The dimension of the \code{mvn} object
}
\description{
Method for obtaining the dimension of an \code{mvn} object.
}
