% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/normal.R
\name{dim.normal}
\alias{dim.normal}
\title{Method for obtaining the dimension of a \code{normal} object.}
\usage{
\method{dim}{normal}(x)
}
\arguments{
\item{x}{The \code{normal} object to obtain the dimension of}
}
\value{
The dimension of the \code{normal} object
}
\description{
Method for obtaining the dimension of a \code{normal} object.
}
