% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/poisson_dist.R
\name{dim.poisson_dist}
\alias{dim.poisson_dist}
\title{Dimension of a Poisson distribution (always 1).}
\usage{
\method{dim}{poisson_dist}(x)
}
\arguments{
\item{x}{A \code{poisson_dist} object.}
}
\value{
\code{1}.
}
\description{
Dimension of a Poisson distribution (always 1).
}
