% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/uniform_dist.R
\name{dim.uniform_dist}
\alias{dim.uniform_dist}
\title{Dimension of a uniform distribution (always 1).}
\usage{
\method{dim}{uniform_dist}(x)
}
\arguments{
\item{x}{A \code{uniform_dist} object.}
}
\value{
\code{1}.
}
\description{
Dimension of a uniform distribution (always 1).
}
