% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/weibull_dist.R
\name{dim.weibull_dist}
\alias{dim.weibull_dist}
\title{Dimension of a Weibull distribution (always 1).}
\usage{
\method{dim}{weibull_dist}(x)
}
\arguments{
\item{x}{A \code{weibull_dist} object.}
}
\value{
\code{1}.
}
\description{
Dimension of a Weibull distribution (always 1).
}
\examples{
dim(weibull_dist(2, 3))
}
