% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/empirical_dist.R
\name{expectation.empirical_dist}
\alias{expectation.empirical_dist}
\title{Method for obtaining the expectation of \code{empirical_dist} object \code{x}
under function \code{g}.}
\usage{
\method{expectation}{empirical_dist}(x, g, ...)
}
\arguments{
\item{x}{The distribution object.}

\item{g}{The function to take the expectation of.}

\item{...}{Additional arguments to pass itno function \code{g}.}
}
\description{
Method for obtaining the expectation of \code{empirical_dist} object \code{x}
under function \code{g}.
}
