% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exp_dist.R
\name{expectation.exp_dist}
\alias{expectation.exp_dist}
\title{This is the expectation of a function \code{g} with respect to an
exponential distribution \code{x} of type \code{exp_dist}.}
\usage{
\method{expectation}{exp_dist}(x, g, rate = NULL, ...)
}
\arguments{
\item{x}{The disrtibution object.}

\item{g}{The function to take the expectation of.}

\item{...}{Additional arguments to pass.}
}
\description{
This is the expectation of a function \code{g} with respect to an
exponential distribution \code{x} of type \code{exp_dist}.
}
