% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generic_dist.R
\name{expectation}
\alias{expectation}
\title{Generic method for obtaining the expectation of \code{f} with respect to
\code{x}.}
\usage{
expectation(x, g, ...)
}
\arguments{
\item{x}{The distribution object.}

\item{g}{The function to take the expectation of.}

\item{...}{Additional arguments to pass into \code{g}.}
}
\value{
The expected value of \code{g(x)}.
}
\description{
Generic method for obtaining the expectation of \code{f} with respect to
\code{x}.
}
\examples{
\donttest{
x <- exponential(1)
# E[X] for Exp(1) is 1
expectation(x, function(t) t)
}
}
