% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/univariate_dist.R
\name{expectation.univariate_dist}
\alias{expectation.univariate_dist}
\title{Method for obtaining the expectation of \code{f} with respect to a
\code{univariate_dist} object \code{x}.}
\usage{
\method{expectation}{univariate_dist}(x, g, ..., control = list())
}
\arguments{
\item{x}{The distribution object.}

\item{g}{The function to take the expectation of.}

\item{...}{Additional arguments to pass into \code{g}.}

\item{control}{An (optional) list of control parameters for \code{integrate} or
\code{expectation_data} (if \code{x} is not continuous)}
}
\description{
Assumes the support is a contiguous interval that has operations
for retrieving the lower and upper bounds.
}
