% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/beta_dist.R
\name{format.beta_dist}
\alias{format.beta_dist}
\title{Format a \code{beta_dist} object as a character string.}
\usage{
\method{format}{beta_dist}(x, ...)
}
\arguments{
\item{x}{A \code{beta_dist} object.}

\item{...}{Additional arguments (not used).}
}
\value{
A character string describing the distribution.
}
\description{
Format a \code{beta_dist} object as a character string.
}
\examples{
format(beta_dist(2, 5))
}
