% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chi_squared.R
\name{format.chi_squared}
\alias{format.chi_squared}
\title{Format a \code{chi_squared} object as a character string.}
\usage{
\method{format}{chi_squared}(x, ...)
}
\arguments{
\item{x}{The \code{chi_squared} object}

\item{...}{Additional arguments (not used)}
}
\value{
A character string describing the distribution
}
\description{
Format a \code{chi_squared} object as a character string.
}
\examples{
format(chi_squared(5))
}
