% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/edist.R
\name{format.edist}
\alias{format.edist}
\title{Format method for \code{edist} objects.}
\usage{
\method{format}{edist}(x, ...)
}
\arguments{
\item{x}{The object to format}

\item{...}{Additional arguments (not used)}
}
\value{
A character string
}
\description{
Format method for \code{edist} objects.
}
\examples{
z <- normal(0, 1) * exponential(2)
format(z)
}
