% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/empirical_dist.R
\name{format.empirical_dist}
\alias{format.empirical_dist}
\title{Format method for \code{empirical_dist} objects.}
\usage{
\method{format}{empirical_dist}(x, ...)
}
\arguments{
\item{x}{The object to format}

\item{...}{Additional arguments (not used)}
}
\value{
A character string
}
\description{
Format method for \code{empirical_dist} objects.
}
\examples{
ed <- empirical_dist(c(1, 2, 3, 4, 5))
format(ed)
}
