% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exponential.R
\name{format.exponential}
\alias{format.exponential}
\title{Format method for \code{exponential} objects.}
\usage{
\method{format}{exponential}(x, ...)
}
\arguments{
\item{x}{The \code{exponential} object to format}

\item{...}{Additional arguments (not used)}
}
\value{
A character string
}
\description{
Format method for \code{exponential} objects.
}
