% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gamma_dist.R
\name{format.gamma_dist}
\alias{format.gamma_dist}
\title{Format a \code{gamma_dist} object as a character string.}
\usage{
\method{format}{gamma_dist}(x, ...)
}
\arguments{
\item{x}{The \code{gamma_dist} object}

\item{...}{Additional arguments (not used)}
}
\value{
A character string describing the distribution
}
\description{
Format a \code{gamma_dist} object as a character string.
}
