% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lognormal.R
\name{format.lognormal}
\alias{format.lognormal}
\title{Format a \code{lognormal} object as a character string.}
\usage{
\method{format}{lognormal}(x, ...)
}
\arguments{
\item{x}{A \code{lognormal} object.}

\item{...}{Additional arguments (not used).}
}
\value{
A character string describing the distribution.
}
\description{
Format a \code{lognormal} object as a character string.
}
\examples{
format(lognormal(0, 1))
}
