% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mvn.R
\name{format.mvn}
\alias{format.mvn}
\title{Format method for \code{mvn} objects.}
\usage{
\method{format}{mvn}(x, ...)
}
\arguments{
\item{x}{The object to format}

\item{...}{Additional arguments (not used)}
}
\value{
A character string
}
\description{
Format method for \code{mvn} objects.
}
