% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/normal.R
\name{format.normal}
\alias{format.normal}
\title{Format method for \code{normal} objects.}
\usage{
\method{format}{normal}(x, ...)
}
\arguments{
\item{x}{The object to format}

\item{...}{Additional arguments (not used)}
}
\value{
A character string
}
\description{
Format method for \code{normal} objects.
}
