% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/poisson_dist.R
\name{format.poisson_dist}
\alias{format.poisson_dist}
\title{Format a \code{poisson_dist} object as a character string.}
\usage{
\method{format}{poisson_dist}(x, ...)
}
\arguments{
\item{x}{A \code{poisson_dist} object.}

\item{...}{Additional arguments (not used).}
}
\value{
A character string describing the distribution.
}
\description{
Format a \code{poisson_dist} object as a character string.
}
\examples{
format(poisson_dist(5))
}
