% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/realized_dist.R
\name{format.realized_dist}
\alias{format.realized_dist}
\title{Format a \code{realized_dist} object as a character string.}
\usage{
\method{format}{realized_dist}(x, ...)
}
\arguments{
\item{x}{A \code{realized_dist} object.}

\item{...}{Additional arguments (not used).}
}
\value{
A character string.
}
\description{
Shows the number of samples and a summary of the source distribution.
}
\examples{
\donttest{
rd <- realize(normal(0, 1), n = 100)
format(rd)
}
}
