% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/uniform_dist.R
\name{format.uniform_dist}
\alias{format.uniform_dist}
\title{Format a \code{uniform_dist} object as a character string.}
\usage{
\method{format}{uniform_dist}(x, ...)
}
\arguments{
\item{x}{A \code{uniform_dist} object.}

\item{...}{Additional arguments (not used).}
}
\value{
A character string describing the distribution.
}
\description{
Format a \code{uniform_dist} object as a character string.
}
\examples{
format(uniform_dist(0, 10))
}
