% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/weibull_dist.R
\name{format.weibull_dist}
\alias{format.weibull_dist}
\title{Format a \code{weibull_dist} object as a character string.}
\usage{
\method{format}{weibull_dist}(x, ...)
}
\arguments{
\item{x}{A \code{weibull_dist} object.}

\item{...}{Additional arguments (not used).}
}
\value{
A character string describing the distribution.
}
\description{
Format a \code{weibull_dist} object as a character string.
}
\examples{
format(weibull_dist(2, 3))
}
