% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gamma_dist.R
\name{gamma_dist}
\alias{gamma_dist}
\title{Construct a gamma distribution object.}
\usage{
gamma_dist(shape, rate)
}
\arguments{
\item{shape}{Shape parameter (positive scalar)}

\item{rate}{Rate parameter (positive scalar)}
}
\value{
A \code{gamma_dist} object
}
\description{
Construct a gamma distribution object.
}
\examples{
x <- gamma_dist(shape = 2, rate = 1)
mean(x)
vcov(x)
format(x)
}
