% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/finite_set.R
\name{has.finite_set}
\alias{has.finite_set}
\title{Determine if a value is contained in the finite set.}
\usage{
\method{has}{finite_set}(object, x)
}
\arguments{
\item{object}{A finite set.}

\item{x}{A vector of values.}
}
\description{
Determine if a value is contained in the finite set.
}
