% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interval.R
\name{has.interval}
\alias{has.interval}
\title{Determine if a value is contained in the interval.}
\usage{
\method{has}{interval}(object, x)
}
\arguments{
\item{object}{An interval object.}

\item{x}{A vector of values.}
}
\value{
Logical vector indicating containment.
}
\description{
Determine if a value is contained in the interval.
}
\examples{
iv <- interval$new(lower = 0, upper = 1)
has(iv, 0.5) # TRUE
has(iv, 2.0) # FALSE
}
