% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/support.R
\name{has}
\alias{has}
\title{Support}
\usage{
has(object, x)
}
\arguments{
\item{object}{A support object.}

\item{x}{A vector of values.}
}
\value{
Logical vector indicating membership.
}
\description{
support is a class that represents the support of a random
element or distribution, i.e. the set of values that it realize.

It's a conceptual class. To satisfy the concept of
a support, the following methods must be implemented:
\enumerate{
\item has: a function that returns a logical vector indicating
whether each value in a vector is contained in the support
\item infimum: a function that returns the infimum of the support
\item supremum: a function that returns the supremum of the support
\item dim: a function that returns the dimension of the support
}

We provide two implementations that satisfy the concept:
\itemize{
\item \code{interval}: a support that is an infiite set of contiguous numeric values
\item \code{finite_set}: a support that is a finite set of values
Determine if a value is contained in the support.
}
}
\examples{
I <- interval$new(0, 1, lower_closed = TRUE, upper_closed = TRUE)
has(I, 0.5)  # TRUE
has(I, 2)    # FALSE

S <- finite_set$new(c(1, 2, 3))
has(S, 2)    # TRUE
has(S, 4)    # FALSE
}
