% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chi_squared.R
\name{hazard.chi_squared}
\alias{hazard.chi_squared}
\title{Method for obtaining the hazard function of a \code{chi_squared} object.}
\usage{
\method{hazard}{chi_squared}(x, ...)
}
\arguments{
\item{x}{The \code{chi_squared} object}

\item{...}{Additional arguments (not used)}
}
\value{
A function that computes h(t) = f(t) / S(t)
}
\description{
Method for obtaining the hazard function of a \code{chi_squared} object.
}
