% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exp_dist.R
\name{hazard.exp_dist}
\alias{hazard.exp_dist}
\title{Method to obtain the hazard function of an \code{exp_dist} object.}
\usage{
\method{hazard}{exp_dist}(x)
}
\arguments{
\item{x}{The \code{exp_dist} object to obtain the hazard function of}
}
\value{
A function that computes the hazard function of the
exponential distribution at a given point \code{t} and rate \code{rate}.
By default, \code{rate} is the failure rate of object \code{x}
Also accepts a \code{log} argument that determines whether
to compute the log of the hazard function.
}
\description{
Method to obtain the hazard function of an \code{exp_dist} object.
}
