% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exponential.R
\name{hazard.exponential}
\alias{hazard.exponential}
\title{Method to obtain the hazard function of an \code{exponential} object.}
\usage{
\method{hazard}{exponential}(x, ...)
}
\arguments{
\item{x}{The \code{exponential} object to obtain the hazard function of}

\item{...}{Additional arguments (not used)}
}
\value{
A function \code{function(t, log = FALSE, ...)} that computes
the hazard rate (or log-hazard) of the exponential distribution.
}
\description{
Method to obtain the hazard function of an \code{exponential} object.
}
\examples{
x <- exponential(rate = 2)
h <- hazard(x)
h(1)
h(5)
}
