% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gamma_dist.R
\name{hazard.gamma_dist}
\alias{hazard.gamma_dist}
\title{Method for obtaining the hazard function of a \code{gamma_dist} object.}
\usage{
\method{hazard}{gamma_dist}(x, ...)
}
\arguments{
\item{x}{The \code{gamma_dist} object}

\item{...}{Additional arguments (not used)}
}
\value{
A function that computes h(t) = f(t) / S(t)
}
\description{
Method for obtaining the hazard function of a \code{gamma_dist} object.
}
\examples{
x <- gamma_dist(shape = 2, rate = 1)
h <- hazard(x)
h(1)
}
