% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/countable_set.R
\name{infimum.countable_set}
\alias{infimum.countable_set}
\title{Get the infimum of a countable set.}
\usage{
\method{infimum}{countable_set}(object)
}
\arguments{
\item{object}{A \code{countable_set} object.}
}
\value{
The lower bound (integer).
}
\description{
Get the infimum of a countable set.
}
\examples{
cs <- countable_set$new(0L)
infimum(cs)  # 0
}
