% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/finite_set.R
\name{infimum.finite_set}
\alias{infimum.finite_set}
\title{Return the infimum of the finite set.}
\usage{
\method{infimum}{finite_set}(object)
}
\arguments{
\item{object}{A finite set.}
}
\description{
Return the infimum of the finite set.
}
