% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interval.R
\name{infimum.interval}
\alias{infimum.interval}
\title{Return the (vector of) infimum of the interval.}
\usage{
\method{infimum}{interval}(object)
}
\arguments{
\item{object}{An interval object.}
}
\value{
Numeric vector of lower bounds.
}
\description{
Return the (vector of) infimum of the interval.
}
\examples{
iv <- interval$new(lower = 0, upper = 1)
infimum(iv) # 0
}
