% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chi_squared.R
\name{inv_cdf.chi_squared}
\alias{inv_cdf.chi_squared}
\title{Method for obtaining the inverse cdf (quantile function) of a \code{chi_squared}
object.}
\usage{
\method{inv_cdf}{chi_squared}(x, ...)
}
\arguments{
\item{x}{The \code{chi_squared} object}

\item{...}{Additional arguments (not used)}
}
\value{
A function that computes the quantile at probability \code{p}
}
\description{
Method for obtaining the inverse cdf (quantile function) of a \code{chi_squared}
object.
}
\examples{
x <- chi_squared(5)
q <- inv_cdf(x)
q(0.5)
q(0.95)
}
