% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exp_dist.R
\name{inv_cdf.exp_dist}
\alias{inv_cdf.exp_dist}
\title{Method to obtain the inverse cdf of an \code{exp_dist} object.}
\usage{
\method{inv_cdf}{exp_dist}(x)
}
\arguments{
\item{x}{The object to obtain the inverse cdf of}
}
\value{
A function that computes the inverse cdf of the exponential
distribution. Accepts as input a vector \code{p} probabilities
to compute the inverse cdf, a \code{rate} value denoting the
failure rate of the exponential distribution, and a logical
\code{log.p} indicating whether input \code{p} denotes probability
or log-probability. By default, \code{rate} is the failure rate
of object \code{x}.
}
\description{
Method to obtain the inverse cdf of an \code{exp_dist} object.
}
