% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exponential.R
\name{inv_cdf.exponential}
\alias{inv_cdf.exponential}
\title{Method to obtain the inverse cdf of an \code{exponential} object.}
\usage{
\method{inv_cdf}{exponential}(x, ...)
}
\arguments{
\item{x}{The object to obtain the inverse cdf of}

\item{...}{Additional arguments (not used)}
}
\value{
A function \code{function(p, lower.tail = TRUE, log.p = FALSE, ...)}
that computes the inverse cdf of the exponential distribution.
}
\description{
Method to obtain the inverse cdf of an \code{exponential} object.
}
\examples{
x <- exponential(rate = 1)
q <- inv_cdf(x)
q(0.5)
q(0.95)
}
