% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/normal.R, R/univariate_dist.R
\name{inv_cdf.normal}
\alias{inv_cdf.normal}
\title{Method for obtaining the inverse cdf of an \code{normal} object.}
\usage{
\method{inv_cdf}{normal}(x, t0 = NULL, eps = 0.001, ...)

\method{inv_cdf}{normal}(x, t0 = NULL, eps = 0.001, ...)
}
\arguments{
\item{x}{The object to obtain the inverse cdf of.}

\item{...}{Additional arguments to pass.}
}
\value{
A function that computes the inverse cdf of the normal distribution.
}
\description{
We use Newton's method to solve for the root of \code{cdf(x)(t) - p}.
}
