% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/beta_dist.R
\name{is_beta_dist}
\alias{is_beta_dist}
\title{Test whether an object is a \code{beta_dist}.}
\usage{
is_beta_dist(x)
}
\arguments{
\item{x}{The object to test.}
}
\value{
\code{TRUE} if \code{x} inherits from \code{"beta_dist"},
\code{FALSE} otherwise.
}
\description{
Test whether an object is a \code{beta_dist}.
}
