% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chi_squared.R
\name{is_chi_squared}
\alias{is_chi_squared}
\title{Test whether an object is a \code{chi_squared}.}
\usage{
is_chi_squared(x)
}
\arguments{
\item{x}{The object to test}
}
\value{
Logical; TRUE if \code{x} inherits from \code{chi_squared}
}
\description{
Test whether an object is a \code{chi_squared}.
}
\examples{
is_chi_squared(chi_squared(3))
is_chi_squared(normal(0, 1))
}
