% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dist.R
\name{is_dist}
\alias{is_dist}
\title{Function to determine whether an object \code{x} is a \code{dist} object.}
\usage{
is_dist(x)
}
\arguments{
\item{x}{The object to test}
}
\description{
Function to determine whether an object \code{x} is a \code{dist} object.
}
