% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/edist.R
\name{is_edist}
\alias{is_edist}
\title{Function to determine whether an object \code{x} is an \code{edist} object.}
\usage{
is_edist(x)
}
\arguments{
\item{x}{The object to test}
}
\value{
Logical; \code{TRUE} if \code{x} is an \code{edist}.
}
\description{
Function to determine whether an object \code{x} is an \code{edist} object.
}
\examples{
is_edist(normal(0, 1) * exponential(1))  # TRUE
is_edist(normal(0, 1))                   # FALSE
}
