% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/empirical_dist.R
\name{is_empirical_dist}
\alias{is_empirical_dist}
\title{Function to determine whether an object \code{x} is an \code{empirical_dist} object.}
\usage{
is_empirical_dist(x)
}
\arguments{
\item{x}{The object to test}
}
\value{
Logical; \code{TRUE} if \code{x} is an \code{empirical_dist}.
}
\description{
Function to determine whether an object \code{x} is an \code{empirical_dist} object.
}
\examples{
ed <- empirical_dist(c(1, 2, 3))
is_empirical_dist(ed)    # TRUE
is_empirical_dist("abc") # FALSE
}
