% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exp_dist.R
\name{is_exp_dist}
\alias{is_exp_dist}
\title{Function to determine whether an object \code{x} is an \code{exp_dist} object.}
\usage{
is_exp_dist(x)
}
\description{
Function to determine whether an object \code{x} is an \code{exp_dist} object.
}
