% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exponential.R
\name{is_exponential}
\alias{is_exponential}
\title{Function to determine whether an object \code{x} is an \code{exponential} object.}
\usage{
is_exponential(x)
}
\arguments{
\item{x}{The object to test}
}
\description{
Function to determine whether an object \code{x} is an \code{exponential} object.
}
