% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gamma_dist.R
\name{is_gamma_dist}
\alias{is_gamma_dist}
\title{Test whether an object is a \code{gamma_dist}.}
\usage{
is_gamma_dist(x)
}
\arguments{
\item{x}{The object to test}
}
\value{
Logical; TRUE if \code{x} inherits from \code{gamma_dist}
}
\description{
Test whether an object is a \code{gamma_dist}.
}
