% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lognormal.R
\name{is_lognormal}
\alias{is_lognormal}
\title{Test whether an object is a \code{lognormal}.}
\usage{
is_lognormal(x)
}
\arguments{
\item{x}{The object to test.}
}
\value{
\code{TRUE} if \code{x} inherits from \code{"lognormal"},
\code{FALSE} otherwise.
}
\description{
Test whether an object is a \code{lognormal}.
}
