% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mixture.R
\name{is_mixture}
\alias{is_mixture}
\title{Test whether an object is a \code{mixture} distribution.}
\usage{
is_mixture(x)
}
\arguments{
\item{x}{The object to test.}
}
\value{
\code{TRUE} if \code{x} inherits from \code{"mixture"},
\code{FALSE} otherwise.
}
\description{
Test whether an object is a \code{mixture} distribution.
}
