% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mvn.R
\name{is_mvn}
\alias{is_mvn}
\title{Function to determine whether an object \code{x} is an \code{mvn} object.}
\usage{
is_mvn(x)
}
\arguments{
\item{x}{The object to test}
}
\description{
Function to determine whether an object \code{x} is an \code{mvn} object.
}
