% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/normal.R
\name{is_normal}
\alias{is_normal}
\title{Function to determine whether an object \code{x} is an \code{normal} object.}
\usage{
is_normal(x)
}
\arguments{
\item{x}{The object to test}
}
\description{
Function to determine whether an object \code{x} is an \code{normal} object.
}
