% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/poisson_dist.R
\name{is_poisson_dist}
\alias{is_poisson_dist}
\title{Test whether an object is a \code{poisson_dist}.}
\usage{
is_poisson_dist(x)
}
\arguments{
\item{x}{The object to test.}
}
\value{
\code{TRUE} if \code{x} inherits from \code{"poisson_dist"},
\code{FALSE} otherwise.
}
\description{
Test whether an object is a \code{poisson_dist}.
}
