% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/realized_dist.R
\name{is_realized_dist}
\alias{is_realized_dist}
\title{Test whether an object is a \code{realized_dist}.}
\usage{
is_realized_dist(x)
}
\arguments{
\item{x}{The object to test.}
}
\value{
\code{TRUE} if \code{x} inherits from \code{"realized_dist"},
\code{FALSE} otherwise.
}
\description{
Test whether an object is a \code{realized_dist}.
}
\examples{
\donttest{
rd <- realize(normal(0, 1), n = 100)
is_realized_dist(rd)  # TRUE
}
is_realized_dist(normal(0, 1))  # FALSE
}
