% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/uniform_dist.R
\name{is_uniform_dist}
\alias{is_uniform_dist}
\title{Test whether an object is a \code{uniform_dist}.}
\usage{
is_uniform_dist(x)
}
\arguments{
\item{x}{The object to test.}
}
\value{
\code{TRUE} if \code{x} inherits from \code{"uniform_dist"},
\code{FALSE} otherwise.
}
\description{
Test whether an object is a \code{uniform_dist}.
}
\examples{
is_uniform_dist(uniform_dist(0, 1))
is_uniform_dist(normal(0, 1))
}
