% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/weibull_dist.R
\name{is_weibull_dist}
\alias{is_weibull_dist}
\title{Test whether an object is a \code{weibull_dist}.}
\usage{
is_weibull_dist(x)
}
\arguments{
\item{x}{The object to test.}
}
\value{
\code{TRUE} if \code{x} inherits from \code{"weibull_dist"},
\code{FALSE} otherwise.
}
\description{
Test whether an object is a \code{weibull_dist}.
}
\examples{
is_weibull_dist(weibull_dist(2, 3))
is_weibull_dist(normal(0, 1))
}
