% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/algebra.R
\name{limit.edist}
\alias{limit.edist}
\title{We have an edist object, which is a subclass of dist, and now we're placing
it inside of a limit expression, where the limit is understood to be with
respect to sample size. We need to define a method for this.}
\usage{
limit.edist(x)
}
\arguments{
\item{x}{The edist object to take the limit of}
}
\value{
The limit of the edist object
}
\description{
We have an edist object, which is a subclass of dist, and now we're placing
it inside of a limit expression, where the limit is understood to be with
respect to sample size. We need to define a method for this.
}
\keyword{internal}
