% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dfr_dist.R
\name{loglik.dfr_dist}
\alias{loglik.dfr_dist}
\title{Log-likelihood contribution of a \code{dfr_dist} object.}
\usage{
loglik.dfr_dist(x, ...)
}
\arguments{
\item{x}{The \code{dfr_dist} object to compute the log-likelihood contribution
for a vector of exact failure times \code{t}.}
}
\description{
Log-likelihood contribution of a \code{dfr_dist} object.
}
