% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mvn.R
\name{marginal.mvn}
\alias{marginal.mvn}
\title{Generic method for obtaining the marginal distribution of an \code{mvn} object
\code{x} over components \code{indices}.}
\usage{
\method{marginal}{mvn}(x, indices)
}
\arguments{
\item{x}{The \code{mvn} object.}

\item{indices}{The indices of the marginal distribution to obtain.}
}
\description{
Generic method for obtaining the marginal distribution of an \code{mvn} object
\code{x} over components \code{indices}.
}
