% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generic_dist.R
\name{marginal}
\alias{marginal}
\title{Generic method for obtaining the marginal distribution of a distribution
object \code{x} over components \code{indices}.}
\usage{
marginal(x, indices)
}
\arguments{
\item{x}{The distribution object.}

\item{indices}{The indices of the marginal distribution to obtain.}
}
\description{
Generic method for obtaining the marginal distribution of a distribution
object \code{x} over components \code{indices}.
}
