% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/beta_dist.R
\name{mean.beta_dist}
\alias{mean.beta_dist}
\title{Mean of a beta distribution.}
\usage{
\method{mean}{beta_dist}(x, ...)
}
\arguments{
\item{x}{A \code{beta_dist} object.}

\item{...}{Additional arguments (not used).}
}
\value{
The mean of the distribution.
}
\description{
Computes \eqn{\alpha / (\alpha + \beta)} where \eqn{\alpha} = \code{shape1}
and \eqn{\beta} = \code{shape2}.
}
\examples{
mean(beta_dist(2, 5))
}
