% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chi_squared.R
\name{mean.chi_squared}
\alias{mean.chi_squared}
\title{Retrieve the mean of a \code{chi_squared} object.}
\usage{
\method{mean}{chi_squared}(x, ...)
}
\arguments{
\item{x}{The \code{chi_squared} object}

\item{...}{Additional arguments (not used)}
}
\value{
The mean, equal to \code{df}
}
\description{
Retrieve the mean of a \code{chi_squared} object.
}
\examples{
mean(chi_squared(10))
}
