% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/edist.R
\name{mean.edist}
\alias{mean.edist}
\title{Method for obtaining the mean of an \code{edist} object.}
\usage{
\method{mean}{edist}(x, n = 10000, ...)
}
\arguments{
\item{x}{The \code{edist} object to retrieve the mean from}

\item{n}{The number of samples to take (default: 10000)}

\item{...}{Additional arguments to pass (not used)}
}
\value{
The mean of the \code{edist} object
}
\description{
Method for obtaining the mean of an \code{edist} object.
}
