% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exp_dist.R
\name{mean.exp_dist}
\alias{mean.exp_dist}
\title{Method to obtain the mean of an \code{exp_dist} object.}
\usage{
\method{mean}{exp_dist}(x, rate = NULL)
}
\arguments{
\item{x}{The \code{exp_dist} object to obtain the mean of}

\item{par}{The rate of the exponential distribution.
If NULL, then the rate of the \code{exp_dist} object
\code{x} is used. We view this as the ability to
override its rate, for instance, when we are
performing estimation and the true rate is unknown.}
}
\description{
Method to obtain the mean of an \code{exp_dist} object.
}
