% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exponential.R
\name{mean.exponential}
\alias{mean.exponential}
\title{Method to obtain the mean of an \code{exponential} object.}
\usage{
\method{mean}{exponential}(x, ...)
}
\arguments{
\item{x}{The \code{exponential} object to obtain the mean of}

\item{...}{Additional arguments (not used)}
}
\description{
Method to obtain the mean of an \code{exponential} object.
}
