% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gamma_dist.R
\name{mean.gamma_dist}
\alias{mean.gamma_dist}
\title{Retrieve the mean of a \code{gamma_dist} object.}
\usage{
\method{mean}{gamma_dist}(x, ...)
}
\arguments{
\item{x}{The \code{gamma_dist} object}

\item{...}{Additional arguments (not used)}
}
\value{
The mean, \code{shape / rate}
}
\description{
Retrieve the mean of a \code{gamma_dist} object.
}
