% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lognormal.R
\name{mean.lognormal}
\alias{mean.lognormal}
\title{Mean of a log-normal distribution.}
\usage{
\method{mean}{lognormal}(x, ...)
}
\arguments{
\item{x}{A \code{lognormal} object.}

\item{...}{Additional arguments (not used).}
}
\value{
The mean of the distribution.
}
\description{
Computes \eqn{\exp(meanlog + sdlog^2 / 2)}.
}
