% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/normal.R
\name{mean.normal}
\alias{mean.normal}
\title{Retrieve the mean of a \code{normal} object.}
\usage{
\method{mean}{normal}(object)
}
\arguments{
\item{object}{The \code{normal} object to retrieve the mean from}
}
\value{
The mean of the \code{normal} object
}
\description{
Retrieve the mean of a \code{normal} object.
}
