% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/poisson_dist.R
\name{mean.poisson_dist}
\alias{mean.poisson_dist}
\title{Mean of a Poisson distribution.}
\usage{
\method{mean}{poisson_dist}(x, ...)
}
\arguments{
\item{x}{A \code{poisson_dist} object.}

\item{...}{Additional arguments (not used).}
}
\value{
The mean, equal to \code{lambda}.
}
\description{
Mean of a Poisson distribution.
}
