% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/uniform_dist.R
\name{mean.uniform_dist}
\alias{mean.uniform_dist}
\title{Mean of a uniform distribution.}
\usage{
\method{mean}{uniform_dist}(x, ...)
}
\arguments{
\item{x}{A \code{uniform_dist} object.}

\item{...}{Additional arguments (not used).}
}
\value{
The mean of the distribution.
}
\description{
Computes \eqn{(min + max) / 2}.
}
\examples{
mean(uniform_dist(0, 10))
}
