% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/weibull_dist.R
\name{mean.weibull_dist}
\alias{mean.weibull_dist}
\title{Mean of a Weibull distribution.}
\usage{
\method{mean}{weibull_dist}(x, ...)
}
\arguments{
\item{x}{A \code{weibull_dist} object.}

\item{...}{Additional arguments (not used).}
}
\value{
The mean of the distribution.
}
\description{
Computes \eqn{scale \cdot \Gamma(1 + 1/shape)}.
}
\examples{
mean(weibull_dist(shape = 2, scale = 3))
}
