% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generic_dist.R
\name{nobs}
\alias{nobs}
\title{This is related to \code{obs}, but in many cases the observations are not stored,
and instead the number of observations is stored. This function retrieves
the number of observations used.}
\usage{
nobs(x)
}
\arguments{
\item{x}{the object to retrieve the number of observations from}
}
\description{
This is related to \code{obs}, but in many cases the observations are not stored,
and instead the number of observations is stored. This function retrieves
the number of observations used.
}
