% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/normal.R
\name{normal}
\alias{normal}
\title{Construct univariate normal distribution object.}
\usage{
normal(mu = 0, var = 1)
}
\arguments{
\item{mu}{mean}

\item{var}{variance}
}
\value{
A \code{normal} distribution object.
}
\description{
Construct univariate normal distribution object.
}
\examples{
x <- normal(mu = 0, var = 1)
mean(x)
vcov(x)
format(x)
}
