% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mixture.R
\name{nparams.mixture}
\alias{nparams.mixture}
\title{Number of parameters for a \code{mixture} distribution.}
\usage{
\method{nparams}{mixture}(x)
}
\arguments{
\item{x}{A \code{mixture} object.}
}
\value{
An integer count of parameters.
}
\description{
The total number of parameters is the sum of component parameters
plus the number of mixing weights.
}
