% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generic_dist.R
\name{nparams}
\alias{nparams}
\title{Generic method for obtaining the number of parameters of
distribution-like object \code{x}.}
\usage{
nparams(x)
}
\arguments{
\item{x}{the object to obtain the number of parameters for}
}
\value{
Integer; the number of parameters.
}
\description{
Generic method for obtaining the number of parameters of
distribution-like object \code{x}.
}
\examples{
d <- empirical_dist(matrix(rnorm(30), ncol = 3))
nparams(d)  # 0 (non-parametric)
}
