% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/empirical_dist.R
\name{obs.empirical_dist}
\alias{obs.empirical_dist}
\title{Method for obtaining the data of a \code{empirical_dist} object.}
\usage{
\method{obs}{empirical_dist}(x, ...)
}
\arguments{
\item{x}{The object to obtain the data of.}

\item{...}{Additional arguments to pass (not used).}
}
\description{
Method for obtaining the data of a \code{empirical_dist} object.
}
