% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generic_dist.R
\name{obs}
\alias{obs}
\title{Generic method for obtaining the observations of an object.}
\usage{
obs(x, ...)
}
\arguments{
\item{x}{The object to obtain the observations of.}

\item{...}{Additional arguments to pass.}
}
\value{
A matrix of observations.
}
\description{
Note that the method \code{nobs} handles the number of observations,
and since \code{obs} may be NULL, then \code{nobs} should be used instead if you
want to know the number of observations. Also, \code{nobs} is a generic method
defined in \code{stats}.
}
