% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/beta_dist.R
\name{params.beta_dist}
\alias{params.beta_dist}
\title{Retrieve the parameters of a \code{beta_dist} object.}
\usage{
\method{params}{beta_dist}(x)
}
\arguments{
\item{x}{A \code{beta_dist} object.}
}
\value{
A named numeric vector with elements \code{shape1} and \code{shape2}.
}
\description{
Retrieve the parameters of a \code{beta_dist} object.
}
