% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chi_squared.R
\name{params.chi_squared}
\alias{params.chi_squared}
\title{Method for obtaining the parameters of a \code{chi_squared} object.}
\usage{
\method{params}{chi_squared}(x)
}
\arguments{
\item{x}{The \code{chi_squared} object}
}
\value{
A named numeric vector of parameters
}
\description{
Method for obtaining the parameters of a \code{chi_squared} object.
}
