% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/edist.R
\name{params.edist}
\alias{params.edist}
\title{Method for obtaining the parameters of an \code{edist} object.}
\usage{
\method{params}{edist}(x)
}
\arguments{
\item{x}{The object to obtain the parameters of}
}
\value{
A named vector of parameters
}
\description{
Method for obtaining the parameters of an \code{edist} object.
}
