% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/empirical_dist.R
\name{params.empirical_dist}
\alias{params.empirical_dist}
\title{\code{empirical_dist} objects have no parameters, so this function returns NULL.}
\usage{
\method{params}{empirical_dist}(x)
}
\arguments{
\item{x}{The empirical distribution object.}
}
\description{
\code{empirical_dist} objects have no parameters, so this function returns NULL.
}
