% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exp_dist.R
\name{params.exp_dist}
\alias{params.exp_dist}
\title{Method for obtaining the parameters of an \code{exp_dist} object.}
\usage{
\method{params}{exp_dist}(x, rate = NULL)
}
\arguments{
\item{x}{The object to obtain the parameters of}
}
\value{
A named vector of parameters
}
\description{
Method for obtaining the parameters of an \code{exp_dist} object.
}
