% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exponential.R
\name{params.exponential}
\alias{params.exponential}
\title{Method for obtaining the parameters of an \code{exponential} object.}
\usage{
\method{params}{exponential}(x)
}
\arguments{
\item{x}{The object to obtain the parameters of}
}
\value{
A named vector of parameters
}
\description{
Method for obtaining the parameters of an \code{exponential} object.
}
\examples{
x <- exponential(rate = 0.5)
params(x)
}
