% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gamma_dist.R
\name{params.gamma_dist}
\alias{params.gamma_dist}
\title{Method for obtaining the parameters of a \code{gamma_dist} object.}
\usage{
\method{params}{gamma_dist}(x)
}
\arguments{
\item{x}{The \code{gamma_dist} object}
}
\value{
A named numeric vector of parameters
}
\description{
Method for obtaining the parameters of a \code{gamma_dist} object.
}
