% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lognormal.R
\name{params.lognormal}
\alias{params.lognormal}
\title{Retrieve the parameters of a \code{lognormal} object.}
\usage{
\method{params}{lognormal}(x)
}
\arguments{
\item{x}{A \code{lognormal} object.}
}
\value{
A named numeric vector with elements \code{meanlog} and \code{sdlog}.
}
\description{
Retrieve the parameters of a \code{lognormal} object.
}
\examples{
params(lognormal(0, 1))
}
