% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mixture.R
\name{params.mixture}
\alias{params.mixture}
\title{Retrieve the parameters of a \code{mixture} object.}
\usage{
\method{params}{mixture}(x)
}
\arguments{
\item{x}{A \code{mixture} object.}
}
\value{
A named numeric vector.
}
\description{
Returns a named numeric vector containing all component parameters
(flattened) followed by the mixing weights.
}
\examples{
m <- mixture(list(normal(0, 1), normal(5, 2)), c(0.3, 0.7))
params(m)
}
