% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mvn.R
\name{params.mvn}
\alias{params.mvn}
\title{Method for obtaining the parameters of a \code{mvn} object.}
\usage{
\method{params}{mvn}(x, ...)
}
\arguments{
\item{x}{The object to obtain the parameters of}

\item{...}{Additional arguments to pass (not used)}
}
\value{
A named vector of parameters
}
\description{
Method for obtaining the parameters of a \code{mvn} object.
}
