% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/poisson_dist.R
\name{params.poisson_dist}
\alias{params.poisson_dist}
\title{Retrieve the parameters of a \code{poisson_dist} object.}
\usage{
\method{params}{poisson_dist}(x)
}
\arguments{
\item{x}{A \code{poisson_dist} object.}
}
\value{
A named numeric vector with element \code{lambda}.
}
\description{
Retrieve the parameters of a \code{poisson_dist} object.
}
\examples{
params(poisson_dist(5))
}
