% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generic_dist.R
\name{params}
\alias{params}
\title{Generic method for obtaining the parameters of an object.}
\usage{
params(x)
}
\arguments{
\item{x}{The object to obtain the parameters of.}
}
\value{
A named vector (or list) of distribution parameters.
}
\description{
Generic method for obtaining the parameters of an object.
}
\examples{
x <- normal(5, 2)
params(x)  # mu = 5, var = 2

y <- exponential(3)
params(y)  # rate = 3
}
