% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/uniform_dist.R
\name{params.uniform_dist}
\alias{params.uniform_dist}
\title{Retrieve the parameters of a \code{uniform_dist} object.}
\usage{
\method{params}{uniform_dist}(x)
}
\arguments{
\item{x}{A \code{uniform_dist} object.}
}
\value{
A named numeric vector with elements \code{min} and \code{max}.
}
\description{
Retrieve the parameters of a \code{uniform_dist} object.
}
\examples{
params(uniform_dist(0, 10))
}
