% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/weibull_dist.R
\name{params.weibull_dist}
\alias{params.weibull_dist}
\title{Retrieve the parameters of a \code{weibull_dist} object.}
\usage{
\method{params}{weibull_dist}(x)
}
\arguments{
\item{x}{A \code{weibull_dist} object.}
}
\value{
A named numeric vector with elements \code{shape} and \code{scale}.
}
\description{
Retrieve the parameters of a \code{weibull_dist} object.
}
\examples{
params(weibull_dist(2, 3))
}
