% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exp_dist.R
\name{pdf.exp_dist}
\alias{pdf.exp_dist}
\title{Method to obtain the pdf of an \code{exp_dist} object.}
\usage{
\method{pdf}{exp_dist}(x)
}
\arguments{
\item{x}{The object to obtain the pdf of}

\item{rate}{The rate of the exponential distribution.}
}
\value{
A function that computes the pdf of the exponential distribution
at a given point \code{t}. Accepts a \code{log} argument that determines
whether to compute the log of the pdf.
}
\description{
Method to obtain the pdf of an \code{exp_dist} object.
}
