% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mvn.R
\name{pdf.mvn}
\alias{pdf.mvn}
\title{Method for obtaining the pdf of an \code{mvn} object.}
\usage{
\method{pdf}{mvn}(x, ...)
}
\arguments{
\item{x}{The object to obtain the pdf of}

\item{...}{Additional arguments to pass (not used)}
}
\value{
A function that computes the pdf of the mvn distribution.
It accepts as input a parameter vector \code{x}, a mean vector \code{mu},
a variance-covariance matrix \code{sigma}, and a \code{log} argument
determining whether to compute the log of the pdf. By default,
\code{mu} and \code{sigma} are the mean and variance-covariance matrix
of object \code{x}.
}
\description{
Method for obtaining the pdf of an \code{mvn} object.
}
