% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generic_dist.R
\name{pdf}
\alias{pdf}
\title{Generic method for obtaining the pdf (probability density function) of an object.}
\usage{
pdf(x, ...)
}
\arguments{
\item{x}{The object to obtain the pdf of.}

\item{...}{Additional arguments to pass.}
}
\description{
Generic method for obtaining the pdf (probability density function) of an object.
}
