% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/edist.R
\name{^.dist}
\alias{^.dist}
\title{Power operator for distribution objects.}
\usage{
\method{^}{dist}(x, y)
}
\arguments{
\item{x}{a dist object (base)}

\item{y}{a numeric scalar (exponent)}
}
\value{
A simplified distribution or edist
}
\description{
Power operator for distribution objects.
}
\examples{
# Standard normal squared yields chi-squared(1)
z <- normal(0, 1)^2
z
}
