% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/beta_dist.R
\name{print.beta_dist}
\alias{print.beta_dist}
\title{Print a \code{beta_dist} object.}
\usage{
\method{print}{beta_dist}(x, ...)
}
\arguments{
\item{x}{A \code{beta_dist} object.}

\item{...}{Additional arguments (not used).}
}
\value{
\code{x}, invisibly.
}
\description{
Print a \code{beta_dist} object.
}
