% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chi_squared.R
\name{print.chi_squared}
\alias{print.chi_squared}
\title{Print method for \code{chi_squared} objects.}
\usage{
\method{print}{chi_squared}(x, ...)
}
\arguments{
\item{x}{The \code{chi_squared} object to print}

\item{...}{Additional arguments (not used)}
}
\description{
Print method for \code{chi_squared} objects.
}
