% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dist.R
\name{print.dist}
\alias{print.dist}
\title{Method for printing a \code{dist} object}
\usage{
\method{print}{dist}(x, ...)
}
\arguments{
\item{x}{The object to print}

\item{...}{Additional arguments to pass}
}
\description{
This method handles printing for probability distribution objects.
It defers to the stats package's print method for distance matrix objects.
}
