% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exp_dist.R
\name{print.exp_dist}
\alias{print.exp_dist}
\title{Print method for \code{exp_dist} objects.}
\usage{
\method{print}{exp_dist}(x, ...)
}
\arguments{
\item{x}{The \code{exp_dist} object to print.}

\item{...}{Additional arguments (not used)}
}
\description{
Print method for \code{exp_dist} objects.
}
