% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exponential.R
\name{print.exponential}
\alias{print.exponential}
\title{Print method for \code{exponential} objects.}
\usage{
\method{print}{exponential}(x, ...)
}
\arguments{
\item{x}{The \code{exponential} object to print.}

\item{...}{Additional arguments (not used)}
}
\description{
Print method for \code{exponential} objects.
}
