% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gamma_dist.R
\name{print.gamma_dist}
\alias{print.gamma_dist}
\title{Print method for \code{gamma_dist} objects.}
\usage{
\method{print}{gamma_dist}(x, ...)
}
\arguments{
\item{x}{The \code{gamma_dist} object to print}

\item{...}{Additional arguments (not used)}
}
\value{
\code{x}, invisibly.
}
\description{
Print method for \code{gamma_dist} objects.
}
\examples{
print(gamma_dist(2, 1))
}
