% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interval.R
\name{print.interval}
\alias{print.interval}
\title{Print the interval.}
\usage{
\method{print}{interval}(x, ...)
}
\arguments{
\item{x}{An interval object.}

\item{...}{Additional arguments.}
}
\description{
Print the interval.
}
