% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lognormal.R
\name{print.lognormal}
\alias{print.lognormal}
\title{Print a \code{lognormal} object.}
\usage{
\method{print}{lognormal}(x, ...)
}
\arguments{
\item{x}{A \code{lognormal} object.}

\item{...}{Additional arguments (not used).}
}
\value{
\code{x}, invisibly.
}
\description{
Print a \code{lognormal} object.
}
