% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mvn.R
\name{print.mvn}
\alias{print.mvn}
\title{Method for printing an \code{mvn} object.}
\usage{
\method{print}{mvn}(x, ...)
}
\arguments{
\item{x}{The object to print}

\item{...}{Additional arguments to pass to \code{print}}
}
\value{
\code{x}, invisibly.
}
\description{
Method for printing an \code{mvn} object.
}
\examples{
print(mvn(c(0, 0)))
}
