% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/normal.R
\name{print.normal}
\alias{print.normal}
\title{Method for printing a \code{normal} object.}
\usage{
\method{print}{normal}(x, ...)
}
\arguments{
\item{x}{The object to print}
}
\description{
Method for printing a \code{normal} object.
}
