% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/poisson_dist.R
\name{print.poisson_dist}
\alias{print.poisson_dist}
\title{Print a \code{poisson_dist} object.}
\usage{
\method{print}{poisson_dist}(x, ...)
}
\arguments{
\item{x}{A \code{poisson_dist} object.}

\item{...}{Additional arguments (not used).}
}
\value{
\code{x}, invisibly.
}
\description{
Print a \code{poisson_dist} object.
}
