% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/realized_dist.R
\name{print.realized_dist}
\alias{print.realized_dist}
\title{Print a \code{realized_dist} object.}
\usage{
\method{print}{realized_dist}(x, ...)
}
\arguments{
\item{x}{A \code{realized_dist} object.}

\item{...}{Additional arguments (not used).}
}
\value{
\code{x}, invisibly.
}
\description{
Print a \code{realized_dist} object.
}
\examples{
\donttest{
rd <- realize(normal(0, 1), n = 100)
print(rd)
}
}
