% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/uniform_dist.R
\name{print.uniform_dist}
\alias{print.uniform_dist}
\title{Print a \code{uniform_dist} object.}
\usage{
\method{print}{uniform_dist}(x, ...)
}
\arguments{
\item{x}{A \code{uniform_dist} object.}

\item{...}{Additional arguments (not used).}
}
\value{
\code{x}, invisibly.
}
\description{
Print a \code{uniform_dist} object.
}
