% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/weibull_dist.R
\name{print.weibull_dist}
\alias{print.weibull_dist}
\title{Print a \code{weibull_dist} object.}
\usage{
\method{print}{weibull_dist}(x, ...)
}
\arguments{
\item{x}{A \code{weibull_dist} object.}

\item{...}{Additional arguments (not used).}
}
\value{
\code{x}, invisibly.
}
\description{
Print a \code{weibull_dist} object.
}
\examples{
print(weibull_dist(2, 3))
}
