% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generic_dist.R
\name{quantile}
\alias{quantile}
\title{Generic method for obtaining the quantile (inverse cdf) of an object.}
\usage{
quantile(x, ...)
}
\arguments{
\item{x}{The object to obtain the quantile of.}

\item{...}{Additional arguments to pass.}
}
\description{
Generic method for obtaining the quantile (inverse cdf) of an object.
}
