% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/beta_dist.R
\name{sampler.beta_dist}
\alias{sampler.beta_dist}
\title{Sampler for a beta distribution.}
\usage{
\method{sampler}{beta_dist}(x, ...)
}
\arguments{
\item{x}{A \code{beta_dist} object.}

\item{...}{Additional arguments (not used).}
}
\value{
A function \code{function(n = 1, ...)} returning a numeric vector
of length \code{n}.
}
\description{
Returns a function that draws \code{n} independent samples from the
beta distribution.
}
