% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chi_squared.R
\name{sampler.chi_squared}
\alias{sampler.chi_squared}
\title{Method for sampling from a \code{chi_squared} object.}
\usage{
\method{sampler}{chi_squared}(x, ...)
}
\arguments{
\item{x}{The \code{chi_squared} object to sample from}

\item{...}{Additional arguments (not used)}
}
\value{
A function that generates \code{n} samples from the chi-squared
distribution
}
\description{
Method for sampling from a \code{chi_squared} object.
}
\examples{
x <- chi_squared(5)
s <- sampler(x)
set.seed(42)
s(5)
}
