% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dist.R
\name{sampler.default}
\alias{sampler.default}
\title{Sampler for non-dist objects (degenerate distributions).}
\usage{
\method{sampler}{default}(x, ...)
}
\arguments{
\item{x}{The object to sample from}

\item{...}{Additional arguments to pass}
}
\value{
A function that takes n and returns n copies of x
}
\description{
Sampler for non-dist objects (degenerate distributions).
}
