% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exp_dist.R
\name{sampler.exp_dist}
\alias{sampler.exp_dist}
\title{Method to sample from an \code{exp_dist} object.}
\usage{
\method{sampler}{exp_dist}(x)
}
\arguments{
\item{x}{The \code{exp_dist} object to sample from.}
}
\value{
A function that allows sampling from the exponential
distribution. Accepts an argument \code{n} denoting sample
size and \code{par} denoting the failure rate. If \code{par} is
NULL, defaults to the failure rate of \code{x}.
}
\description{
Method to sample from an \code{exp_dist} object.
}
