% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exponential.R
\name{sampler.exponential}
\alias{sampler.exponential}
\title{Method to sample from an \code{exponential} object.}
\usage{
\method{sampler}{exponential}(x, ...)
}
\arguments{
\item{x}{The \code{exponential} object to sample from.}

\item{...}{Additional arguments to pass (not used)}
}
\value{
A function \code{function(n = 1, ...)} that draws \code{n}
samples from the exponential distribution.
}
\description{
Method to sample from an \code{exponential} object.
}
\examples{
x <- exponential(rate = 2)
s <- sampler(x)
set.seed(42)
s(5)
}
