% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gamma_dist.R
\name{sampler.gamma_dist}
\alias{sampler.gamma_dist}
\title{Method for sampling from a \code{gamma_dist} object.}
\usage{
\method{sampler}{gamma_dist}(x, ...)
}
\arguments{
\item{x}{The \code{gamma_dist} object to sample from}

\item{...}{Additional arguments (not used)}
}
\value{
A function that generates \code{n} samples from the gamma distribution
}
\description{
Method for sampling from a \code{gamma_dist} object.
}
\examples{
x <- gamma_dist(shape = 2, rate = 1)
s <- sampler(x)
set.seed(42)
s(5)
}
