% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mvn.R
\name{sampler.mvn}
\alias{sampler.mvn}
\title{Function generator for sampling from a \code{mvn} (multivariate normal) object.}
\usage{
\method{sampler}{mvn}(x, ...)
}
\arguments{
\item{x}{The \code{mvn} object to sample from}

\item{...}{Additional arguments passed to \code{rmvnorm} on every call.}
}
\description{
Function generator for sampling from a \code{mvn} (multivariate normal) object.
}
\examples{
X <- mvn(c(0, 0), diag(2))
s <- sampler(X)
set.seed(42)
s(3)
}
