% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/normal.R
\name{sampler.normal}
\alias{sampler.normal}
\title{Method for sampling from a \code{normal} object.}
\usage{
\method{sampler}{normal}(x, ...)
}
\arguments{
\item{x}{The \code{normal} object to sample from}

\item{...}{Additional arguments to pass (not used)}
}
\value{
A function \code{function(n = 1, ...)} that draws \code{n}
samples from the normal distribution.
}
\description{
Method for sampling from a \code{normal} object.
}
\examples{
x <- normal(0, 1)
s <- sampler(x)
set.seed(42)
s(5)
}
