% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/poisson_dist.R
\name{sampler.poisson_dist}
\alias{sampler.poisson_dist}
\title{Sampler for a Poisson distribution.}
\usage{
\method{sampler}{poisson_dist}(x, ...)
}
\arguments{
\item{x}{A \code{poisson_dist} object.}

\item{...}{Additional arguments (not used).}
}
\value{
A function \code{function(n = 1, ...)} returning an integer vector
of length \code{n}.
}
\description{
Returns a function that draws \code{n} independent samples from the
Poisson distribution.
}
