% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generic_dist.R
\name{sampler}
\alias{sampler}
\title{Generic method for obtaining the sampler of an object}
\usage{
sampler(x, ...)
}
\arguments{
\item{x}{The object to obtain the sampler of.}

\item{...}{Additional arguments to pass.}
}
\description{
Generic method for obtaining the sampler of an object
}
