% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/weibull_dist.R
\name{sampler.weibull_dist}
\alias{sampler.weibull_dist}
\title{Sampler for a Weibull distribution.}
\usage{
\method{sampler}{weibull_dist}(x, ...)
}
\arguments{
\item{x}{A \code{weibull_dist} object.}

\item{...}{Additional arguments (not used).}
}
\value{
A function \code{function(n = 1, ...)} returning a numeric vector
of length \code{n}.
}
\description{
Returns a function that draws \code{n} independent samples from the
Weibull distribution.
}
