% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generic_dist.R
\name{sd}
\alias{sd}
\title{Computes the stadnard deviation of a distribution-like object.}
\usage{
sd(x)
}
\arguments{
\item{x}{the distribution-like object to compute the sd of}
}
\description{
Computes the stadnard deviation of a distribution-like object.
}
