% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/algebra.R
\name{simplify.dist}
\alias{simplify.dist}
\title{Default Method for simplifying a \code{dist} object. Just returns the object.}
\usage{
\method{simplify}{dist}(x, ...)
}
\arguments{
\item{x}{The \code{dist} object to simplify}

\item{...}{Additional arguments to pass (not used)}
}
\value{
The \code{dist} object
}
\description{
Default Method for simplifying a \code{dist} object. Just returns the object.
}
\examples{
x <- normal(0, 1)
identical(simplify(x), x)  # TRUE, returns unchanged
}
