% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/edist.R
\name{/.dist}
\alias{/.dist}
\title{Division of distribution objects.}
\usage{
\method{/}{dist}(x, y)
}
\arguments{
\item{x}{first operand}

\item{y}{second operand}
}
\value{
A simplified distribution or edist
}
\description{
Handles dist / scalar (delegates to dist * (1/scalar)),
scalar / dist, and dist / dist.
}
