% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dist.R
\name{summary.dist}
\alias{summary.dist}
\title{Method for obtaining a summary of a \code{dist} object.}
\usage{
\method{summary}{dist}(object, ..., name = NULL, nobs = NULL)
}
\arguments{
\item{object}{The object to obtain the summary of}

\item{...}{Additional arguments to pass}

\item{name}{The name of the distribution, defaults to the class of the object.}

\item{nobs}{The number of observations to report for the summary, if applicable.}
}
\value{
A \code{summary_dist} object
}
\description{
Method for obtaining a summary of a \code{dist} object.
}
\examples{
summary(normal(0, 1))
}
