% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dist.R
\name{summary_dist}
\alias{summary_dist}
\title{Method for constructing a \code{summary_dist} object.}
\usage{
summary_dist(name, mean, vcov, nobs = NULL)
}
\arguments{
\item{name}{The name of the distribution}

\item{mean}{The mean of the distribution}

\item{vcov}{The variance of the distribution}

\item{nobs}{The number of observations used to construct the distribution,
if applicable.}
}
\value{
A \code{summary_dist} object
}
\description{
Method for constructing a \code{summary_dist} object.
}
