% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/beta_dist.R
\name{sup.beta_dist}
\alias{sup.beta_dist}
\title{Support of a beta distribution.}
\usage{
\method{sup}{beta_dist}(x)
}
\arguments{
\item{x}{A \code{beta_dist} object.}
}
\value{
An \code{interval} object representing \eqn{(0, 1)}.
}
\description{
The beta distribution is supported on the open interval \eqn{(0, 1)}.
}
\examples{
sup(beta_dist(2, 5))
}
