% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chi_squared.R
\name{sup.chi_squared}
\alias{sup.chi_squared}
\title{Support for chi-squared distribution, the positive real numbers (0, Inf).}
\usage{
\method{sup}{chi_squared}(x)
}
\arguments{
\item{x}{The \code{chi_squared} object}
}
\value{
An \code{interval} object representing (0, Inf)
}
\description{
Support for chi-squared distribution, the positive real numbers (0, Inf).
}
