% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/edist.R
\name{sup.edist}
\alias{sup.edist}
\title{Support for expression distributions.}
\usage{
\method{sup}{edist}(x)
}
\arguments{
\item{x}{An \code{edist} object.}
}
\value{
A \code{finite_set} support object.
}
\description{
Falls back to \code{\link{realize}} and delegates to
\code{\link{sup.empirical_dist}}.
}
\examples{
\donttest{
set.seed(1)
z <- normal(0, 1) * exponential(1)
sup(z)
}
}
