% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/empirical_dist.R
\name{sup.empirical_dist}
\alias{sup.empirical_dist}
\title{Method for obtaining the support of \code{empirical_dist} object \code{x}.}
\usage{
\method{sup}{empirical_dist}(x)
}
\arguments{
\item{x}{The empirical distribution object.}
}
\value{
A \code{finite_set} object containing the support of \code{x}.
}
\description{
Method for obtaining the support of \code{empirical_dist} object \code{x}.
}
