% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exp_dist.R
\name{sup.exp_dist}
\alias{sup.exp_dist}
\title{Support for exponential distribution, the positive real numbers, (0, Inf).}
\usage{
\method{sup}{exp_dist}(x)
}
\arguments{
\item{x}{The object to obtain the support of}
}
\value{
An \code{interval} object representing the support of the exponential
}
\description{
Support for exponential distribution, the positive real numbers, (0, Inf).
}
