% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exponential.R
\name{sup.exponential}
\alias{sup.exponential}
\title{Support for exponential distribution, the positive real numbers, (0, Inf).}
\usage{
\method{sup}{exponential}(x)
}
\arguments{
\item{x}{The object to obtain the support of}
}
\value{
An \code{interval} object representing the support of the exponential
}
\description{
Support for exponential distribution, the positive real numbers, (0, Inf).
}
\examples{
x <- exponential(rate = 1)
sup(x)
}
