% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gamma_dist.R
\name{sup.gamma_dist}
\alias{sup.gamma_dist}
\title{Support for gamma distribution, the positive real numbers (0, Inf).}
\usage{
\method{sup}{gamma_dist}(x)
}
\arguments{
\item{x}{The \code{gamma_dist} object}
}
\value{
An \code{interval} object representing (0, Inf)
}
\description{
Support for gamma distribution, the positive real numbers (0, Inf).
}
\examples{
sup(gamma_dist(2, 1))
}
