% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lognormal.R
\name{sup.lognormal}
\alias{sup.lognormal}
\title{Support of a log-normal distribution.}
\usage{
\method{sup}{lognormal}(x)
}
\arguments{
\item{x}{A \code{lognormal} object.}
}
\value{
An \code{interval} object representing \eqn{(0, \infty)}.
}
\description{
The log-normal distribution is supported on \eqn{(0, \infty)}.
}
