% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mixture.R
\name{sup.mixture}
\alias{sup.mixture}
\title{Support of a mixture distribution.}
\usage{
\method{sup}{mixture}(x)
}
\arguments{
\item{x}{A \code{mixture} object.}
}
\value{
An \code{interval} object.
}
\description{
Returns an \code{\link{interval}} spanning the widest range of all
component supports (from the smallest infimum to the largest supremum).
}
\examples{
m <- mixture(list(normal(0, 1), exponential(1)), c(0.5, 0.5))
sup(m)
}
