% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mvn.R
\name{sup.mvn}
\alias{sup.mvn}
\title{Method for obtaining the support of a \code{mvn} object, where the support
is defined as values that have non-zero probability density.}
\usage{
\method{sup}{mvn}(x, ...)
}
\arguments{
\item{x}{The \code{mvn} object to obtain the support of}

\item{...}{Additional arguments to pass (not used)}
}
\value{
A support-type object (see \code{support.R}), in this case an
\code{interval} object for each component.
}
\description{
Method for obtaining the support of a \code{mvn} object, where the support
is defined as values that have non-zero probability density.
}
