% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/normal.R
\name{sup.normal}
\alias{sup.normal}
\title{Method for obtaining the support of a \code{normal} object, where the support
is defined as values that have non-zero probability density.}
\usage{
\method{sup}{normal}(x)
}
\arguments{
\item{x}{The \code{normal} object to obtain the support of}
}
\value{
A support-type object (see \code{support.R}), in this case an
\code{interval} object for each component.
}
\description{
Method for obtaining the support of a \code{normal} object, where the support
is defined as values that have non-zero probability density.
}
\examples{
x <- normal(0, 1)
sup(x)
}
