% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/poisson_dist.R
\name{sup.poisson_dist}
\alias{sup.poisson_dist}
\title{Support of a Poisson distribution.}
\usage{
\method{sup}{poisson_dist}(x)
}
\arguments{
\item{x}{A \code{poisson_dist} object.}
}
\value{
A \code{countable_set} object with lower bound 0.
}
\description{
The Poisson distribution is supported on the non-negative integers
\eqn{\{0, 1, 2, \ldots\}}.
}
\examples{
sup(poisson_dist(5))
}
