% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/uniform_dist.R
\name{sup.uniform_dist}
\alias{sup.uniform_dist}
\title{Support of a uniform distribution.}
\usage{
\method{sup}{uniform_dist}(x)
}
\arguments{
\item{x}{A \code{uniform_dist} object.}
}
\value{
An \code{interval} object representing \eqn{[min, max]}.
}
\description{
The uniform distribution is supported on \eqn{[min, max]}.
}
