% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/weibull_dist.R
\name{sup.weibull_dist}
\alias{sup.weibull_dist}
\title{Support of a Weibull distribution.}
\usage{
\method{sup}{weibull_dist}(x)
}
\arguments{
\item{x}{A \code{weibull_dist} object.}
}
\value{
An \code{interval} object representing \eqn{(0, \infty)}.
}
\description{
The Weibull distribution is supported on \eqn{(0, \infty)}.
}
