% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/countable_set.R
\name{supremum.countable_set}
\alias{supremum.countable_set}
\title{Get the supremum of a countable set.}
\usage{
\method{supremum}{countable_set}(object)
}
\arguments{
\item{object}{A \code{countable_set} object.}
}
\value{
\code{Inf} (the set is unbounded above).
}
\description{
Get the supremum of a countable set.
}
