% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/finite_set.R
\name{supremum.finite_set}
\alias{supremum.finite_set}
\title{Return the supremum of the finite set.}
\usage{
\method{supremum}{finite_set}(object)
}
\arguments{
\item{object}{A finite set.}
}
\description{
Return the supremum of the finite set.
}
