% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interval.R
\name{supremum.interval}
\alias{supremum.interval}
\title{Return the (vector of) supremum of the interval.}
\usage{
\method{supremum}{interval}(object)
}
\arguments{
\item{object}{An interval object.}
}
\description{
Return the (vector of) supremum of the interval.
}
