% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chi_squared.R
\name{surv.chi_squared}
\alias{surv.chi_squared}
\title{Method for obtaining the survival function of a \code{chi_squared} object.}
\usage{
\method{surv}{chi_squared}(x, ...)
}
\arguments{
\item{x}{The \code{chi_squared} object}

\item{...}{Additional arguments (not used)}
}
\value{
A function that computes S(t) = P(X > t)
}
\description{
Method for obtaining the survival function of a \code{chi_squared} object.
}
