% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exponential.R
\name{surv.exponential}
\alias{surv.exponential}
\title{Method to obtain the survival function of an \code{exponential} object.}
\usage{
\method{surv}{exponential}(x, ...)
}
\arguments{
\item{x}{The object to obtain the survival function of}

\item{...}{Additional arguments (not used)}
}
\value{
A function \code{function(t, log.p = FALSE, ...)} that computes
the survival function \eqn{S(t) = P(X > t)}.
}
\description{
Method to obtain the survival function of an \code{exponential} object.
}
\examples{
x <- exponential(rate = 1)
S <- surv(x)
S(1)
S(2)
}
