% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gamma_dist.R
\name{surv.gamma_dist}
\alias{surv.gamma_dist}
\title{Method for obtaining the survival function of a \code{gamma_dist} object.}
\usage{
\method{surv}{gamma_dist}(x, ...)
}
\arguments{
\item{x}{The \code{gamma_dist} object}

\item{...}{Additional arguments (not used)}
}
\value{
A function that computes S(t) = P(X > t)
}
\description{
Method for obtaining the survival function of a \code{gamma_dist} object.
}
