% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/normal.R
\name{surv.normal}
\alias{surv.normal}
\title{Method for obtaining the survival function of an \code{normal} object.}
\usage{
\method{surv}{normal}(x)
}
\arguments{
\item{x}{The object to obtain the survival function of}
}
\value{
A function that computes the survival of the normal distribution.
It accepts as input a parameter vector \code{x}, a mean vector \code{mu},
a variance-covariance matrix \code{sigma}, and a \code{log} argument
determining whether to compute the log of the cdf. By default,
\code{mu} and \code{sigma} are the mean and variance of object \code{x}.
}
\description{
Method for obtaining the survival function of an \code{normal} object.
}
