% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dist.R, R/generic_dist.R
\name{surv.continuous_dist}
\alias{surv.continuous_dist}
\alias{surv}
\title{Default survival function for continuous distributions.}
\usage{
\method{surv}{continuous_dist}(x, ...)

surv(x, ...)
}
\arguments{
\item{x}{The object to obtain the survival function of.}

\item{...}{Additional arguments to pass.}
}
\value{
A function \code{function(t, ...)} returning the survival
probability.

A function computing the survival function \eqn{S(t) = P(X > t)}.
}
\description{
Computes \eqn{S(t) = 1 - F(t)} from the CDF.
}
\examples{
x <- beta_dist(2, 5)
S <- surv(x)
S(0.5)
x <- exponential(1)
S <- surv(x)
S(0)  # 1 (survival at time 0)
S(1)  # exp(-1), approximately 0.368
}
