% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/univariate_dist.R
\name{var.univariate_dist}
\alias{var.univariate_dist}
\title{Method for obtaining the variance of \code{univariate_dist} object \code{x}.}
\usage{
var.univariate_dist(x, ...)
}
\arguments{
\item{x}{The distribution object.}

\item{...}{Additional arguments to pass.}
}
\description{
Method for obtaining the variance of \code{univariate_dist} object \code{x}.
}
