% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/beta_dist.R
\name{vcov.beta_dist}
\alias{vcov.beta_dist}
\title{Variance of a beta distribution.}
\usage{
\method{vcov}{beta_dist}(object, ...)
}
\arguments{
\item{object}{A \code{beta_dist} object.}

\item{...}{Additional arguments (not used).}
}
\value{
The variance (scalar).
}
\description{
Computes \eqn{\alpha\beta / ((\alpha+\beta)^2 (\alpha+\beta+1))}.
}
\examples{
vcov(beta_dist(2, 5))
}
