% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chi_squared.R
\name{vcov.chi_squared}
\alias{vcov.chi_squared}
\title{Retrieve the variance of a \code{chi_squared} object.}
\usage{
\method{vcov}{chi_squared}(object, ...)
}
\arguments{
\item{object}{The \code{chi_squared} object}

\item{...}{Additional arguments (not used)}
}
\value{
The variance, \code{2 * df}
}
\description{
Retrieve the variance of a \code{chi_squared} object.
}
\examples{
vcov(chi_squared(10))
}
