% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dist.R
\name{vcov.default}
\alias{vcov.default}
\title{Variance-covariance for non-dist objects (degenerate distributions).}
\usage{
\method{vcov}{default}(object, ...)
}
\arguments{
\item{object}{The object (returns 0 for constants)}

\item{...}{Additional arguments to pass (not used)}
}
\value{
0 (degenerate distributions have no variance)
}
\description{
Variance-covariance for non-dist objects (degenerate distributions).
}
