% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/edist.R
\name{vcov.edist}
\alias{vcov.edist}
\title{Method for obtaining the variance-covariance matrix (or scalar)}
\usage{
\method{vcov}{edist}(x, n = 1000, ...)
}
\arguments{
\item{x}{The \code{edist} object to retrieve the variance-covariance matrix from}

\item{n}{The number of samples to take (default: 1000)}

\item{...}{Additional arguments to pass (not used)}
}
\value{
The variance-covariance matrix of the \code{edist} object
}
\description{
Method for obtaining the variance-covariance matrix (or scalar)
}
