% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/empirical_dist.R
\name{vcov.empirical_dist}
\alias{vcov.empirical_dist}
\title{Method for obtaining the variance of \code{empirical_dist} object \code{x}.}
\usage{
\method{vcov}{empirical_dist}(object, ...)
}
\arguments{
\item{object}{The empirical distribution object.}

\item{...}{Additional arguments to pass (not used).}
}
\value{
The sample variance-covariance matrix.
}
\description{
Method for obtaining the variance of \code{empirical_dist} object \code{x}.
}
\examples{
ed <- empirical_dist(c(1, 2, 3, 4, 5))
vcov(ed) # sample variance

ed_mv <- empirical_dist(matrix(rnorm(20), ncol = 2))
vcov(ed_mv) # 2x2 covariance matrix
}
