% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exp_dist.R
\name{vcov.exp_dist}
\alias{vcov.exp_dist}
\title{Method for obtaining the variance of a \code{exp_dist} object.}
\usage{
\method{vcov}{exp_dist}(object, rate = NULL)
}
\arguments{
\item{object}{The \code{exp_dist} object to obtain the variance of}

\item{par}{The rate parameter of the exponential distribution.
If NULL, then the rate of the \code{exp_dist} object
\code{object} is used. We view this as the ability to
override its rate, for instance, when we are
performing estimation and the true rate is unknown.}
}
\description{
Method for obtaining the variance of a \code{exp_dist} object.
}
