% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exponential.R
\name{vcov.exponential}
\alias{vcov.exponential}
\title{Retrieve the variance of a \code{exponential} object.}
\usage{
\method{vcov}{exponential}(x)
}
\arguments{
\item{x}{The \code{exponential} object to retrieve the variance for}
}
\value{
The variance-covariance matrix of the \code{normal} object
}
\description{
Retrieve the variance of a \code{exponential} object.
}
