% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gamma_dist.R
\name{vcov.gamma_dist}
\alias{vcov.gamma_dist}
\title{Retrieve the variance of a \code{gamma_dist} object.}
\usage{
\method{vcov}{gamma_dist}(object, ...)
}
\arguments{
\item{object}{The \code{gamma_dist} object}

\item{...}{Additional arguments (not used)}
}
\value{
The variance, \code{shape / rate^2}
}
\description{
Retrieve the variance of a \code{gamma_dist} object.
}
