% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lognormal.R
\name{vcov.lognormal}
\alias{vcov.lognormal}
\title{Variance of a log-normal distribution.}
\usage{
\method{vcov}{lognormal}(object, ...)
}
\arguments{
\item{object}{A \code{lognormal} object.}

\item{...}{Additional arguments (not used).}
}
\value{
The variance (scalar).
}
\description{
Computes \eqn{(\exp(sdlog^2) - 1) \exp(2 \cdot meanlog + sdlog^2)}.
}
