% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mvn.R
\name{vcov.mvn}
\alias{vcov.mvn}
\title{Retrieve the variance-covariance matrix of an \code{mvn} object.}
\usage{
\method{vcov}{mvn}(object, ...)
}
\arguments{
\item{object}{The \code{mvn} object to retrieve the variance-covariance matrix of}

\item{...}{Additional arguments to pass (not used)}
}
\value{
The variance-covariance matrix of the \code{mvn} object
}
\description{
Retrieve the variance-covariance matrix of an \code{mvn} object.
}
