% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/normal.R
\name{vcov.normal}
\alias{vcov.normal}
\title{Retrieve the variance-covariance matrix (or scalar)
of a \code{normal} object.}
\usage{
\method{vcov}{normal}(object, ...)
}
\arguments{
\item{object}{The \code{normal} object to retrieve the variance-covariance matrix from}

\item{...}{Additional arguments to pass (not used)}
}
\value{
The variance-covariance matrix of the \code{normal} object
}
\description{
Retrieve the variance-covariance matrix (or scalar)
of a \code{normal} object.
}
\examples{
x <- normal(0, 4)
vcov(x)
}
