% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/poisson_dist.R
\name{vcov.poisson_dist}
\alias{vcov.poisson_dist}
\title{Variance of a Poisson distribution.}
\usage{
\method{vcov}{poisson_dist}(object, ...)
}
\arguments{
\item{object}{A \code{poisson_dist} object.}

\item{...}{Additional arguments (not used).}
}
\value{
The variance (scalar), equal to \code{lambda}.
}
\description{
For the Poisson distribution, the variance equals \code{lambda}.
}
