% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generic_dist.R
\name{vcov}
\alias{vcov}
\title{Computes the variance-covariance matrix of a distribution-like object.}
\usage{
vcov(x)
}
\arguments{
\item{x}{the distribution-like object to compute the variance-covariance}
}
\description{
Computes the variance-covariance matrix of a distribution-like object.
}
