% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/uniform_dist.R
\name{vcov.uniform_dist}
\alias{vcov.uniform_dist}
\title{Variance of a uniform distribution.}
\usage{
\method{vcov}{uniform_dist}(object, ...)
}
\arguments{
\item{object}{A \code{uniform_dist} object.}

\item{...}{Additional arguments (not used).}
}
\value{
The variance (scalar).
}
\description{
Computes \eqn{(max - min)^2 / 12}.
}
\examples{
vcov(uniform_dist(0, 10))
}
