% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/univariate_dist.R
\name{vcov.univariate_dist}
\alias{vcov.univariate_dist}
\title{Method for obtaining the variance of \code{univariate_dist} object \code{x}.}
\usage{
\method{vcov}{univariate_dist}(x)
}
\arguments{
\item{x}{The distribution object.}
}
\description{
Method for obtaining the variance of \code{univariate_dist} object \code{x}.
}
