% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/weibull_dist.R
\name{vcov.weibull_dist}
\alias{vcov.weibull_dist}
\title{Variance of a Weibull distribution.}
\usage{
\method{vcov}{weibull_dist}(object, ...)
}
\arguments{
\item{object}{A \code{weibull_dist} object.}

\item{...}{Additional arguments (not used).}
}
\value{
The variance (scalar).
}
\description{
Computes \eqn{scale^2 (\Gamma(1 + 2/shape) - [\Gamma(1 + 1/shape)]^2)}.
}
\examples{
vcov(weibull_dist(shape = 2, scale = 3))
}
