% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/empirical_dist.R
\name{cdf.empirical_dist}
\alias{cdf.empirical_dist}
\title{Method for obtaining the cdf of \code{empirical_dist} object \code{x}.}
\usage{
\method{cdf}{empirical_dist}(x)
}
\arguments{
\item{x}{The empirical distribution object.}
}
\value{
A function that takes a numeric vector \code{t} and returns the
empirical cdf of \code{x} evaluated at \code{t}.
}
\description{
If \code{x} is a multivariate empirical distribution, this function will
throw an error. It's only defined for univariate empirical distributions.
}
