% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gamma_dist.R
\name{cdf.gamma_dist}
\alias{cdf.gamma_dist}
\title{Method for obtaining the cdf of a \code{gamma_dist} object.}
\usage{
\method{cdf}{gamma_dist}(x, ...)
}
\arguments{
\item{x}{The \code{gamma_dist} object}

\item{...}{Additional arguments (not used)}
}
\value{
A function that computes the cdf at point(s) \code{t}
}
\description{
Method for obtaining the cdf of a \code{gamma_dist} object.
}
\examples{
x <- gamma_dist(shape = 2, rate = 1)
F <- cdf(x)
F(1)
F(2)
}
