% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mvn.R
\name{cdf.mvn}
\alias{cdf.mvn}
\title{Method for obtaining the CDF of a \code{mvn} object.}
\usage{
\method{cdf}{mvn}(x, ...)
}
\arguments{
\item{x}{The object to obtain the CDF of}

\item{...}{Additional arguments to pass (not used)}
}
\value{
A function computing the multivariate normal CDF.
}
\description{
Method for obtaining the CDF of a \code{mvn} object.
}
\examples{
X <- mvn(c(0, 0), diag(2))
F <- cdf(X)
F(c(0, 0))
}
