% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generic_dist.R
\name{cdf}
\alias{cdf}
\title{Generic method for obtaining the cdf of an object.}
\usage{
cdf(x, ...)
}
\arguments{
\item{x}{The object to obtain the cdf of.}

\item{...}{Additional arguments to pass.}
}
\description{
Generic method for obtaining the cdf of an object.
}
