% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generic_dist.R
\name{conditional}
\alias{conditional}
\title{Generic method for obtaining the conditional distribution of a distribution
object \code{x} given condition \code{P}.}
\usage{
conditional(x, P)
}
\arguments{
\item{x}{The empirical distribution object.}

\item{P}{the predicate to condition on.}
}
\description{
Generic method for obtaining the conditional distribution of a distribution
object \code{x} given condition \code{P}.
}
