% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/empirical_dist.R
\name{density.empirical_dist}
\alias{density.empirical_dist}
\title{Method for obtaining the pdf of a \code{empirical_dist} object.}
\usage{
\method{density}{empirical_dist}(x, ...)
}
\arguments{
\item{x}{The object to obtain the pdf of.}

\item{...}{Additional arguments to pass into the pdf function.}
}
\value{
A function computing the empirical PMF at given points.
}
\description{
Method for obtaining the pdf of a \code{empirical_dist} object.
}
\note{
sort tibble lexicographically and do a binary search to find upper
and lower bound in \code{log(nobs(x))} time.
}
\examples{
ed <- empirical_dist(c(1, 2, 2, 3, 3, 3))
f <- density(ed)
f(2)          # 2/6
f(3, log = TRUE) # log(3/6)
}
