% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/empirical_dist.R
\name{dim.empirical_dist}
\alias{dim.empirical_dist}
\title{Method for obtaining the dimension of a \code{empirical_dist} object.}
\usage{
\method{dim}{empirical_dist}(x)
}
\arguments{
\item{x}{The object to obtain the dimension of.}
}
\value{
Integer; the number of dimensions.
}
\description{
Method for obtaining the dimension of a \code{empirical_dist} object.
}
\examples{
ed1 <- empirical_dist(c(1, 2, 3))
dim(ed1) # 1

ed2 <- empirical_dist(matrix(1:6, ncol = 2))
dim(ed2) # 2
}
