% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/finite_set.R
\name{dim.finite_set}
\alias{dim.finite_set}
\title{Return the dimension of the finite set.}
\usage{
\method{dim}{finite_set}(x)
}
\arguments{
\item{x}{A finite set.}
}
\description{
Return the dimension of the finite set.
}
