% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interval.R
\name{dim.interval}
\alias{dim.interval}
\title{Return the dimension of the interval.}
\usage{
\method{dim}{interval}(x)
}
\arguments{
\item{x}{An interval object.}
}
\description{
Return the dimension of the interval.
}
