% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/edist.R
\name{Summary.dist}
\alias{Summary.dist}
\title{Summary group generic for distribution objects.}
\usage{
\method{Summary}{dist}(..., na.rm = FALSE)
}
\arguments{
\item{...}{dist objects}

\item{na.rm}{ignored}
}
\value{
A simplified distribution or edist
}
\description{
Handles sum(), prod(), min(), max() of distributions.
}
