% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/edist.R
\name{edist}
\alias{edist}
\title{Takes an expression \code{e} and a list \code{vars} and returns a
lazy \code{edist} (expression distribution object), that is a subclass
of \code{dist} that can be used in place of a \code{dist} object.}
\usage{
edist(e, vars)
}
\arguments{
\item{e}{the expression to evaluate against the arguments.}

\item{vars}{the list of distributions (with variable names)
to evaluate the expression \code{e} against.}
}
\description{
Takes an expression \code{e} and a list \code{vars} and returns a
lazy \code{edist} (expression distribution object), that is a subclass
of \code{dist} that can be used in place of a \code{dist} object.
}
