% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/empirical_dist.R
\name{empirical_dist}
\alias{empirical_dist}
\title{Construct empirical distribution object.}
\usage{
empirical_dist(data)
}
\arguments{
\item{data}{data to construct empirical distribution from. if matrix or
data frame, each row is a joint observation, if a vector, each
element is an observation. whatever data is, it must be
convertible to a tibble.}
}
\description{
Construct empirical distribution object.
}
