% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dist.R
\name{expectation.dist}
\alias{expectation.dist}
\title{Expectation operator applied to \code{object} of type \code{dist}
with respect to a function \code{g}. That is, \code{E(g(object))}.}
\usage{
\method{expectation}{dist}(object, g = function(t) t, n = 10000, alpha = 0.05)
}
\arguments{
\item{object}{\code{dist} object}

\item{g}{characteristic function of interest, defaults to identity}

\item{n}{number of sample points, larger samples reduce variance}

\item{alpha}{significance level for the confidence interval}
}
\description{
We use the CLT to construct a CI(\code{alpha}) for the estimate of the
expectation. That is, we estimate \code{E(g(object))} with the sample mean
and Var(g(object)) with the sigma^2/n, where sigma^2 is the sample
variance of g(object) and n is the number of samples. From these,
we construct the CI.
}
