% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generic_dist.R
\name{expectation}
\alias{expectation}
\title{Generic method for obtaining the expectation of \code{f} with respect to
\code{x}.}
\usage{
expectation(x, g, ...)
}
\arguments{
\item{x}{The disrtibution object.}

\item{g}{The function to take the expectation of.}

\item{...}{Additional arguments to pass into \code{g}.}
}
\description{
Generic method for obtaining the expectation of \code{f} with respect to
\code{x}.
}
