% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/univariate_dist.R
\name{expectation.univariate_dist}
\alias{expectation.univariate_dist}
\title{Method for obtaining the expectation of \code{f} with respect to a
\code{univariate_dist} object \code{x}.}
\usage{
\method{expectation}{univariate_dist}(x, ..., control = list())
}
\arguments{
\item{x}{The distribution object.}

\item{...}{Additional arguments to pass into \code{pdf} and \code{sup}.}

\item{control}{An (optional) list of control parameters for \code{integrate}.}

\item{g}{The function to take the expectation of.}
}
\description{
Method for obtaining the expectation of \code{f} with respect to a
\code{univariate_dist} object \code{x}.
}
