% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exponential.R
\name{exponential}
\alias{exponential}
\title{Construct exponential distribution object.}
\usage{
exponential(rate)
}
\arguments{
\item{rate}{failure rate}
}
\value{
An \code{exponential} distribution object.
}
\description{
Construct exponential distribution object.
}
\examples{
x <- exponential(rate = 2)
mean(x)
vcov(x)
format(x)
}
