% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mixture.R
\name{format.mixture}
\alias{format.mixture}
\title{Format a \code{mixture} object as a character string.}
\usage{
\method{format}{mixture}(x, ...)
}
\arguments{
\item{x}{A \code{mixture} object.}

\item{...}{Additional arguments (not used).}
}
\value{
A character string describing the mixture.
}
\description{
Format a \code{mixture} object as a character string.
}
