% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/countable_set.R
\name{has.countable_set}
\alias{has.countable_set}
\title{Check membership in a countable set.}
\usage{
\method{has}{countable_set}(object, x)
}
\arguments{
\item{object}{A \code{countable_set} object.}

\item{x}{Value(s) to check.}
}
\description{
Returns \code{TRUE} if all values are integers (within floating-point
tolerance) that are at least as large as the lower bound.
}
