% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/finite_set.R
\name{has.finite_set}
\alias{has.finite_set}
\title{Determine if a value is contained in the finite set.}
\usage{
\method{has}{finite_set}(object, x)
}
\arguments{
\item{object}{A finite set.}

\item{x}{A vector of values.}
}
\value{
Logical indicating membership.
}
\description{
Determine if a value is contained in the finite set.
}
\examples{
fs <- finite_set$new(c(1, 3, 5, 7))
has(fs, 3) # TRUE
has(fs, 4) # FALSE
}
