% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exponential.R
\name{hazard.exponential}
\alias{hazard.exponential}
\title{Method to obtain the hazard function of an \code{exponential} object.}
\usage{
\method{hazard}{exponential}(x, ...)
}
\arguments{
\item{x}{The \code{exponential} object to obtain the hazard function of}

\item{...}{Additional arguments (not used)}
}
\value{
A function that computes the hazard function of the
exponential distribution at a given point \code{t} and rate \code{rate}.
By default, \code{rate} is the failure rate of object \code{x}
Also accepts a \code{log} argument that determines whether
to compute the log of the hazard function.
}
\description{
Method to obtain the hazard function of an \code{exponential} object.
}
