% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generic_dist.R
\name{hazard}
\alias{hazard}
\title{Generic method for obtaining the hazard function of an object.}
\usage{
hazard(x, ...)
}
\arguments{
\item{x}{The object to obtain the hazard function of.}

\item{...}{Additional arguments to pass.}
}
\description{
Generic method for obtaining the hazard function of an object.
}
