% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generic_dist.R
\name{hazard}
\alias{hazard}
\title{Generic method for obtaining the hazard function of an object.}
\usage{
hazard(x, ...)
}
\arguments{
\item{x}{The object to obtain the hazard function of.}

\item{...}{Additional arguments to pass.}
}
\value{
A function computing the hazard rate at given points.
}
\description{
Generic method for obtaining the hazard function of an object.
}
\examples{
x <- exponential(2)
h <- hazard(x)
h(1)  # hazard rate at t = 1 (constant for exponential)
}
