% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exponential.R
\name{inv_cdf.exponential}
\alias{inv_cdf.exponential}
\title{Method to obtain the inverse cdf of an \code{exponential} object.}
\usage{
\method{inv_cdf}{exponential}(x, ...)
}
\arguments{
\item{x}{The object to obtain the inverse cdf of}

\item{...}{Additional arguments (not used)}
}
\value{
A function that computes the inverse cdf of the exponential
distribution. Accepts as input a vector \code{p} probabilities
to compute the inverse cdf, a \code{rate} value denoting the
failure rate of the exponential distribution, and a logical
\code{log.p} indicating whether input \code{p} denotes probability
or log-probability. By default, \code{rate} is the failure rate
of object \code{x}.
}
\description{
Method to obtain the inverse cdf of an \code{exponential} object.
}
\examples{
x <- exponential(rate = 1)
q <- inv_cdf(x)
q(0.5)
q(0.95)
}
