% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gamma_dist.R
\name{inv_cdf.gamma_dist}
\alias{inv_cdf.gamma_dist}
\title{Method for obtaining the inverse cdf (quantile function) of a \code{gamma_dist}
object.}
\usage{
\method{inv_cdf}{gamma_dist}(x, ...)
}
\arguments{
\item{x}{The \code{gamma_dist} object}

\item{...}{Additional arguments (not used)}
}
\value{
A function that computes the quantile at probability \code{p}
}
\description{
Method for obtaining the inverse cdf (quantile function) of a \code{gamma_dist}
object.
}
