% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/normal.R
\name{inv_cdf.normal}
\alias{inv_cdf.normal}
\title{Method for obtaining the inverse cdf of an \code{normal} object.}
\usage{
\method{inv_cdf}{normal}(x, ...)
}
\arguments{
\item{x}{The object to obtain the inverse cdf of}

\item{...}{Additional arguments to pass (not used)}
}
\value{
A function that computes the inverse cdf of the normal distribution.
}
\description{
Method for obtaining the inverse cdf of an \code{normal} object.
}
