% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generic_dist.R
\name{inv_cdf}
\alias{inv_cdf}
\title{Generic method for obtaining the quantile (inverse cdf) of an object.}
\usage{
inv_cdf(x, ...)
}
\arguments{
\item{x}{The object to obtain the quantile of.}

\item{...}{Additional arguments to pass.}
}
\description{
Generic method for obtaining the quantile (inverse cdf) of an object.
}
