% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/edist.R
\name{is_edist}
\alias{is_edist}
\title{Function to determine whether an object \code{x} is an \code{edist} object.}
\usage{
is_edist(x)
}
\arguments{
\item{x}{The object to test}
}
\description{
Function to determine whether an object \code{x} is an \code{edist} object.
}
