% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mvn.R
\name{is_mvn}
\alias{is_mvn}
\title{Function to determine whether an object \code{x} is an \code{mvn} object.}
\usage{
is_mvn(x)
}
\arguments{
\item{x}{The object to test}
}
\value{
Logical; \code{TRUE} if \code{x} is an \code{mvn}.
}
\description{
Function to determine whether an object \code{x} is an \code{mvn} object.
}
\examples{
is_mvn(mvn(c(0, 0)))
is_mvn(normal(0, 1))
}
