% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/empirical_dist.R
\name{marginal.empirical_dist}
\alias{marginal.empirical_dist}
\title{Method for obtaining the marginal distribution of \code{empirical_dist} object
\code{x}.}
\usage{
\method{marginal}{empirical_dist}(x, indices)
}
\arguments{
\item{x}{The empirical distribution object.}

\item{indices}{The indices of the marginal distribution to obtain.}
}
\description{
Method for obtaining the marginal distribution of \code{empirical_dist} object
\code{x}.
}
