% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/edist.R
\name{Math.dist}
\alias{Math.dist}
\title{Math group generic for distribution objects.}
\usage{
\method{Math}{dist}(x, ...)
}
\arguments{
\item{x}{a dist object}

\item{...}{additional arguments}
}
\value{
A simplified distribution or edist
}
\description{
Handles exp(), log(), sqrt(), abs(), cos(), sin(), etc.
}
