% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mvn.R
\name{mean.mvn}
\alias{mean.mvn}
\title{Retrieve the mean of a \code{mvn} object.}
\usage{
\method{mean}{mvn}(x, ...)
}
\arguments{
\item{x}{The \code{mvn} object to retrieve the mean from}

\item{...}{Additional arguments to pass (not used)}
}
\value{
The mean of the \code{mvn} object
}
\description{
Retrieve the mean of a \code{mvn} object.
}
