% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/univariate_dist.R
\name{mean.univariate_dist}
\alias{mean.univariate_dist}
\title{Method for obtaining the mean of \code{univariate_dist} object \code{x}.}
\usage{
\method{mean}{univariate_dist}(x, ...)
}
\arguments{
\item{x}{The distribution object.}

\item{...}{Additional arguments to pass into \code{expectation}.}
}
\description{
Method for obtaining the mean of \code{univariate_dist} object \code{x}.
}
