% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mvn.R
\name{mvn}
\alias{mvn}
\title{Construct (multivariate or univariate) normal distribution object.}
\usage{
mvn(mu = 0, sigma = diag(length(mu)))
}
\arguments{
\item{mu}{mean}

\item{sigma}{variance-covariance matrix}
}
\description{
Construct (multivariate or univariate) normal distribution object.
}
