% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/empirical_dist.R
\name{nobs.empirical_dist}
\alias{nobs.empirical_dist}
\title{Method for obtaining the number of observations in a \code{empirical_dist}
object.}
\usage{
\method{nobs}{empirical_dist}(object, ...)
}
\arguments{
\item{object}{The object to obtain the number of observations of.}

\item{...}{Additional arguments to pass (not used).}
}
\value{
The number of observations in \code{x}.
}
\description{
Method for obtaining the number of observations in a \code{empirical_dist}
object.
}
