% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/empirical_dist.R
\name{nobs.empirical_dist}
\alias{nobs.empirical_dist}
\title{Method for obtaining the number of observations used to construct a
\code{empirical_dist} object.}
\usage{
\method{nobs}{empirical_dist}(object, ...)
}
\arguments{
\item{object}{The empirical distribution object.}

\item{...}{Additional arguments to pass (not used).}
}
\description{
Method for obtaining the number of observations used to construct a
\code{empirical_dist} object.
}
