% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/normal.R
\name{normal}
\alias{normal}
\title{Construct univariate normal distribution object.}
\usage{
normal(mu = 0, var = 1)
}
\arguments{
\item{mu}{mean}

\item{var}{variance}
}
\description{
Construct univariate normal distribution object.
}
