% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generic_dist.R
\name{nparams}
\alias{nparams}
\title{Generic method for obtaining the number of parameters of
distribution-like object \code{x}.}
\usage{
nparams(x)
}
\arguments{
\item{x}{the object to obtain the number of parameters for}
}
\description{
Generic method for obtaining the number of parameters of
distribution-like object \code{x}.
}
