% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/empirical_dist.R
\name{obs.empirical_dist}
\alias{obs.empirical_dist}
\title{Method for obtaining the observations used to construct a
\code{empirical_dist} object.}
\usage{
\method{obs}{empirical_dist}(x)
}
\arguments{
\item{x}{The empirical distribution object.}
}
\description{
Method for obtaining the observations used to construct a
\code{empirical_dist} object.
}
