% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generic_dist.R
\name{params}
\alias{params}
\title{Generic method for obtaining the parameters of an object.}
\usage{
params(x, ...)
}
\arguments{
\item{x}{The object to obtain the parameters of.}

\item{...}{Additional arguments to pass.}
}
\description{
Generic method for obtaining the parameters of an object.
}
