% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/empirical_dist.R
\name{pdf.empirical_dist}
\alias{pdf.empirical_dist}
\title{Method for obtaining the pdf of a \code{empirical_dist} object.}
\usage{
\method{pdf}{empirical_dist}(x)
}
\arguments{
\item{x}{The object to obtain the pdf of.}
}
\description{
Method for obtaining the pdf of a \code{empirical_dist} object.
}
\note{
sort tibble lexographically and do a binary search to find upper
and lower bound in \code{log(nobs(x))} time.
}
