% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/edist.R
\name{print.edist}
\alias{print.edist}
\title{Method for printing an \code{edist} object.}
\usage{
\method{print}{edist}(x, ...)
}
\arguments{
\item{x}{The object to print}

\item{...}{Additional arguments to pass (not used)}
}
\description{
Method for printing an \code{edist} object.
}
\examples{
edist(expression(x + y),
     list("x" = normal(mu = 0, var = 1),
          "y" = exponential(rate = 1)))
#> Distribution x + y
#>   x ~ normal(mu = 0, var = 1)
#>   y ~ exponential(rate = 1)
}
