% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/edist.R
\name{print.edist}
\alias{print.edist}
\title{Print method for \code{edist} objects.}
\usage{
\method{print}{edist}(x, ...)
}
\arguments{
\item{x}{The object to print}

\item{...}{Additional arguments to pass (not used)}
}
\description{
Print method for \code{edist} objects.
}
