% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/empirical_dist.R
\name{print.empirical_dist}
\alias{print.empirical_dist}
\title{Print method for \code{empirical_dist} objects.}
\usage{
\method{print}{empirical_dist}(x, ...)
}
\arguments{
\item{x}{The object to print}

\item{...}{Additional arguments to pass}
}
\description{
Print method for \code{empirical_dist} objects.
}
