% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mixture.R
\name{print.mixture}
\alias{print.mixture}
\title{Print a \code{mixture} object.}
\usage{
\method{print}{mixture}(x, ...)
}
\arguments{
\item{x}{A \code{mixture} object.}

\item{...}{Additional arguments (not used).}
}
\value{
\code{x}, invisibly.
}
\description{
Print a \code{mixture} object.
}
\examples{
m <- mixture(list(normal(0, 1), normal(5, 1)), c(0.5, 0.5))
print(m)
}
