% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/normal.R
\name{print.normal}
\alias{print.normal}
\title{Print method for \code{normal} objects.}
\usage{
\method{print}{normal}(x, ...)
}
\arguments{
\item{x}{The object to print}

\item{...}{Additional arguments to pass (not used)}
}
\description{
Print method for \code{normal} objects.
}
