% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dist.R
\name{print.summary_dist}
\alias{print.summary_dist}
\title{Print method for \code{summary_dist} objects.}
\usage{
\method{print}{summary_dist}(x, ...)
}
\arguments{
\item{x}{The object to print}

\item{...}{Additional arguments}
}
\value{
\code{x}, invisibly.
}
\description{
Print method for \code{summary_dist} objects.
}
\examples{
s <- summary(normal(5, 2))
print(s)
}
