% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/empirical_dist.R
\name{rmap.empirical_dist}
\alias{rmap.empirical_dist}
\title{Method for obtaining the empirical distribution of a function of the
observations of \code{empirical_dist} object \code{x}.}
\usage{
\method{rmap}{empirical_dist}(x, g, ...)
}
\arguments{
\item{x}{The empirical distribution object.}

\item{g}{The function to apply to each observation.}

\item{...}{Additional arguments to pass into function \code{g}.}
}
\value{
An \code{empirical_dist} of the transformed observations.
}
\description{
Method for obtaining the empirical distribution of a function of the
observations of \code{empirical_dist} object \code{x}.
}
\examples{
ed <- empirical_dist(c(1, 2, 3, 4))
ed2 <- rmap(ed, function(x) x^2)
mean(ed2) # mean of 1, 4, 9, 16
}
