% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mvn.R
\name{rmap.mvn}
\alias{rmap.mvn}
\title{Computes the distribution of \code{g(x)} where \code{x} is an \code{mvn} object.}
\usage{
\method{rmap}{mvn}(x, g, n = 10000L, ...)
}
\arguments{
\item{x}{The distribution object.}

\item{g}{The function to apply.}

\item{n}{number of samples to take to estimate distribution of \code{g(x)} if
\code{method} is \code{mc} or \code{empirical}. Defaults to 10000.}

\item{...}{additional arguments to pass into the \code{g} function.}
}
\description{
By the invariance property, if \code{x} is an \code{mvn} object,
then under the right conditions, asymptotically, \code{g(x)} is an MVN
distributed,
g(x) ~ normal(g(mean(x)), sigma)
where \code{sigma} is the variance-covariance of \code{g(x)}
}
\details{
We provide two different methods for estimating the
variance-covariance of \code{g(x)}:
method = "delta" -> delta method
method = "mc" -> monte carlo method

Finally, we provide an all-purpose \code{empirical} method, where we just sample
from the MVN and place the sample in an \code{empirical_dist} object, and then
apply the \code{rmap} with \code{g} to that.
}
