% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generic_dist.R
\name{rmap}
\alias{rmap}
\title{Generic method for applying a map \code{f} to distribution object \code{x}.}
\usage{
rmap(x, g, ...)
}
\arguments{
\item{x}{The distribution object.}

\item{g}{The function to apply.}

\item{...}{Additional arguments to pass.}
}
\description{
Generic method for applying a map \code{f} to distribution object \code{x}.
}
