% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mvn.R
\name{sample_mvn_region}
\alias{sample_mvn_region}
\title{Function for obtaining sample points for an \code{mvn} object that is within
the \code{p}-probability region.}
\usage{
sample_mvn_region(n, mu, sigma, p = 0.95, ...)
}
\arguments{
\item{n}{the sample size}

\item{mu}{mean vector}

\item{sigma}{variance-covariance matrix}

\item{p}{the probability region}

\item{...}{additional arguments to pass into \code{mahalanobis}}
}
\description{
Function for obtaining sample points for an \code{mvn} object that is within
the \code{p}-probability region.
}
