% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/empirical_dist.R
\name{sampler.empirical_dist}
\alias{sampler.empirical_dist}
\title{Method for obtaining the sampler for a \code{empirical_dist} object.}
\usage{
\method{sampler}{empirical_dist}(x, ...)
}
\arguments{
\item{x}{The object to obtain the sampler of.}

\item{...}{Additional arguments to pass (not used).}
}
\value{
A function that takes \code{n} and returns \code{n} resampled
observations.
}
\description{
Method for obtaining the sampler for a \code{empirical_dist} object.
}
\examples{
ed <- empirical_dist(c(10, 20, 30))
s <- sampler(ed)
set.seed(42)
s(5)
}
