% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lognormal.R
\name{sampler.lognormal}
\alias{sampler.lognormal}
\title{Sampler for a log-normal distribution.}
\usage{
\method{sampler}{lognormal}(x, ...)
}
\arguments{
\item{x}{A \code{lognormal} object.}

\item{...}{Additional arguments (not used).}
}
\value{
A function \code{function(n = 1, ...)} returning a numeric vector
of length \code{n}.
}
\description{
Returns a function that draws \code{n} independent samples from the
log-normal distribution.
}
