% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mvn.R
\name{sampler.mvn}
\alias{sampler.mvn}
\title{Method for sampling from a \code{mvn} (multivariate normal) object.}
\usage{
\method{sampler}{mvn}(x, ...)
}
\arguments{
\item{x}{The \code{mvn} object to sample from}

\item{`...`}{Additional arguments to pass to \code{rmvnorm} or \code{sample_mvn_region}}
}
\description{
Method for sampling from a \code{mvn} (multivariate normal) object.
}
