% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/normal.R
\name{sampler.normal}
\alias{sampler.normal}
\title{Method for sampling from a \code{normal} object.}
\usage{
\method{sampler}{normal}(x)
}
\arguments{
\item{x}{The \code{normal} object to sample from}
}
\value{
A function that samples from the normal distribution. As input,
it accepts a sample size \code{n}, a numeric \code{mu}, and a variance
numeric \code{var}. By default, \code{mu} and \code{var} are the mean and
variance of object \code{x}.
}
\description{
Method for sampling from a \code{normal} object.
}
