% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/finite_set.R
\name{supremum.finite_set}
\alias{supremum.finite_set}
\title{Return the supremum of the finite set.}
\usage{
\method{supremum}{finite_set}(object)
}
\arguments{
\item{object}{A finite set.}
}
\value{
Numeric; the maximum value(s).
}
\description{
Return the supremum of the finite set.
}
\examples{
fs <- finite_set$new(c(1, 3, 5, 7))
supremum(fs) # 7
}
