% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interval.R
\name{supremum.interval}
\alias{supremum.interval}
\title{Return the (vector of) supremum of the interval.}
\usage{
\method{supremum}{interval}(object)
}
\arguments{
\item{object}{An interval object.}
}
\value{
Numeric vector of upper bounds.
}
\description{
Return the (vector of) supremum of the interval.
}
\examples{
iv <- interval$new(lower = 0, upper = 1)
supremum(iv) # 1
}
