% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generic_dist.R
\name{surv}
\alias{surv}
\title{Generic method for obtaining the survival function of an object.}
\usage{
surv(x, ...)
}
\arguments{
\item{x}{The object to obtain the survival function of.}

\item{...}{Additional arguments to pass.}
}
\description{
Generic method for obtaining the survival function of an object.
}
