% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/empirical_dist.R
\name{vcov.empirical_dist}
\alias{vcov.empirical_dist}
\title{Method for obtaining the variance of \code{empirical_dist} object \code{x}.}
\usage{
\method{vcov}{empirical_dist}(x, ...)
}
\arguments{
\item{x}{The empirical distribution object.}

\item{...}{Additional arguments to pass (not used)}
}
\description{
Method for obtaining the variance of \code{empirical_dist} object \code{x}.
}
