% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mvn.R
\name{vcov.mvn}
\alias{vcov.mvn}
\title{Retrieve the variance-covariance matrix (or scalar)
of a \code{mvn} object.
A variance-covariance matrix is a square matrix
giving the covariance between each pair of elements
of a given random vector. Importantly, its diagonal
contains variances of the elements}
\usage{
\method{vcov}{mvn}(object, ...)
}
\arguments{
\item{object}{The \code{mvn} object to retrieve
the variance-covariance matrix from}

\item{...}{Additional arguments to pass (not used)}
}
\value{
The variance-covariance matrix of the \code{mvn} object
}
\description{
Retrieve the variance-covariance matrix (or scalar)
of a \code{mvn} object.
A variance-covariance matrix is a square matrix
giving the covariance between each pair of elements
of a given random vector. Importantly, its diagonal
contains variances of the elements
}
