% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/normal.R
\name{vcov.normal}
\alias{vcov.normal}
\title{Retrieve the variance-covariance matrix (or scalar)
of a \code{normal} object.
A variance-covariance matrix is a square matrix
giving the covariance between each pair of elements
of a given random vector. Importantly, its diagonal
contains variances of the elements}
\usage{
\method{vcov}{normal}(object)
}
\arguments{
\item{object}{The \code{normal} object to retrieve
the variance-covariance matrix from}
}
\value{
The variance-covariance matrix of the \code{normal} object
}
\description{
Retrieve the variance-covariance matrix (or scalar)
of a \code{normal} object.
A variance-covariance matrix is a square matrix
giving the covariance between each pair of elements
of a given random vector. Importantly, its diagonal
contains variances of the elements
}
